refactor(app): move tab.rs's four sidecars and its tests into tab/
`tab.rs` carried one `impl Tab` block of 3,700 lines, so a history-paging
bug was read through the canvas painter and the canvas through the paging
state machine. Four groups that a `Tab` *does* — rather than what it *is* —
move to sibling modules under `crates/app/src/tab/`, the way `app/` and
`pane/` already hold theirs, and the four inline test modules move to
`tab/tests/mod.rs` under their existing names.

Bodies are unchanged. A line multiset over the old file against the new
five shows the only altered lines are seven `pub(super)` widenings, one
call qualified as `quantick_feed::demo_gap_ms()` because `mod feed` now
occupies the name `feed` in this module, and the import that alias left.

`OlderCandles` and `CanvasChrome` keep their `crate::tab::` paths through
`pub use`, so no file outside `tab*` changes an import.

tab.rs 4,472 -> 1,661 production lines; the ceiling is tightened and the
size `!budget` falls 50,996 -> 48,185.
