+++ crates/app/src/tab.rs
@@ -42,10 +42,18 @@ use quantick_feed::history_reach::{
 };
 use quantick_feed::stall::{self, Stall, StallInput};
 use quantick_feed::{
-    self as feed, FeedCommand, FeedConnectionState, FeedEvent, FeedGap, FeedHandle, FeedLatency,
-    FeedNotice, MAX_REMEMBERED_GAPS, MIN_MARKED_GAP_MS, ReplayLink, past_resume_floor,
+    FeedCommand, FeedConnectionState, FeedEvent, FeedGap, FeedHandle, FeedLatency, FeedNotice,
+    MAX_REMEMBERED_GAPS, MIN_MARKED_GAP_MS, ReplayLink, past_resume_floor,
 };
 use std::path::PathBuf;
+
+mod canvas;
+mod feed;
+mod history;
+mod strategies;
+
+pub use canvas::CanvasChrome;
+pub use history::OlderCandles;
 
 /// Each UI capture epoch reserves room for reconnect generations. This keeps
 /// late events from an aborted task below the next accepted generation floor.
@@ -155,170 +163,6 @@ impl From<CanvasLayout> for crate::config::DeclaredLayout {
             CanvasLayout::Time => crate::config::DeclaredLayout::Time,
             CanvasLayout::TimeAndFlow => crate::config::DeclaredLayout::TimeAndFlow,
             CanvasLayout::TimeTimeAndFlow => crate::config::DeclaredLayout::TimeTimeAndFlow,
-        }
-    }
-}
-
-/// The window chrome a tab's canvas borrows for one frame. The tab completes
-/// it with its own symbol to make the [`PaneChrome`] its panes read.
-pub struct CanvasChrome<'a> {
-    pub toolrail: &'a mut ToolRail,
-    pub presets: &'a crate::drawings::presets::PresetStore,
-    /// See [`PaneChrome::begin_text_edit`].
-    pub begin_text_edit: &'a mut bool,
-    pub style: &'a ChartStyle,
-    pub tz: TzOffset,
-    /// What the running source can produce, for the layer menu's disabled
-    /// entries. Resolved once by the app rather than per pane, per entry.
-    pub capabilities: FeedCapabilities,
-    /// Whether the source infers the aggressor side (see
-    /// [`PaneChrome::side_inferred`]). Resolved once, like `capabilities`.
-    pub side_inferred: bool,
-    /// The footprint layer's signal tunables (see [`PaneChrome::footprint`]).
-    pub footprint: &'a mut crate::footprint_config::FootprintConfig,
-    /// Where a pane's layer menu leaves the switches it does not own.
-    pub layers: &'a mut crate::chart_layers::LayerActions,
-}
-
-/// Put an older slice of venue candles in front of the ones already held,
-/// keeping the base ascending by `open_time` and free of duplicates.
-///
-/// The fast path is the one progressive loading actually produces: the slice
-/// is strictly older than everything held, so it is spliced in front and the
-/// order is already right. The merge below exists for the case the port
-/// permits but no provider aims for — a window that overlaps what is held,
-/// through a venue re-reporting a bucket at a boundary. There the candle
-/// already on screen wins: it is the one the trader has been reading, and a
-/// bar that redraws itself for no visible reason is worse than a bar fetched
-/// a second apart from an identical twin.
-fn merge_older_candles(base: &mut Vec<quantick_engine::Bar>, older: Vec<quantick_engine::Bar>) {
-    let disjoint = match (older.last(), base.first()) {
-        (Some(newest_incoming), Some(oldest_held)) => {
-            newest_incoming.open_time < oldest_held.open_time
-        }
-        _ => true,
-    };
-    if disjoint {
-        base.splice(0..0, older);
-        return;
-    }
-    let mut merged: std::collections::BTreeMap<i64, quantick_engine::Bar> =
-        older.into_iter().map(|bar| (bar.open_time, bar)).collect();
-    for bar in base.drain(..) {
-        merged.insert(bar.open_time, bar);
-    }
-    *base = merged.into_values().collect();
-}
-
-/// Drop venue bars that overlap the trade-derived series.
-///
-/// The two series meet at a seam, and the composed chart is only searchable if
-/// `open_time` never decreases across it. A venue candle covering the same
-/// window as the first engine bar would sit *after* it in time while sitting
-/// before it in slot order, so every venue bucket from that one on is dropped:
-/// what the app cut from prints is the better record of that window anyway.
-///
-/// With no engine bars yet the whole prefix stands — there is nothing to
-/// overlap.
-fn trim_to_seam(
-    mut folded: Vec<quantick_engine::Bar>,
-    first_engine_bar: Option<&quantick_engine::Bar>,
-    partial: Option<&quantick_engine::Bar>,
-    interval_ms: i64,
-) -> Vec<quantick_engine::Bar> {
-    let Some(seam) = seam_bucket_ms(first_engine_bar, partial, interval_ms) else {
-        return folded;
-    };
-    folded.retain(|bar| bar.open_time < seam);
-    folded
-}
-
-/// The same trim over a block the caller only has on loan.
-///
-/// Two functions rather than one taking a `Cow`, because they pay for
-/// different things and both paths are on the diet. The owning one above trims
-/// a vector `resample::fold` just built and is about to drop — a `retain` there
-/// copies nothing at all. This one is handed the venue's whole base, which the
-/// tab keeps, so it copies out only the bars that survive rather than cloning a
-/// week of minutes in order to throw most of them away.
-fn trim_borrowed_to_seam(
-    base: &[quantick_engine::Bar],
-    first_engine_bar: Option<&quantick_engine::Bar>,
-    partial: Option<&quantick_engine::Bar>,
-    interval_ms: i64,
-) -> Vec<quantick_engine::Bar> {
-    let Some(seam) = seam_bucket_ms(first_engine_bar, partial, interval_ms) else {
-        return base.to_vec();
-    };
-    base.iter()
-        .filter(|bar| bar.open_time < seam)
-        .cloned()
-        .collect()
-}
-
-/// Where the venue's candles have to stop for the pane's own bars to begin, or
-/// `None` when there are no bars yet and the whole block stands.
-///
-/// Buckets, not stamps. A venue candle's `open_time` is its bucket start; an
-/// engine bar's is its *first trade*, which sits strictly inside the bucket.
-/// Comparing the two raw would keep the venue candle covering the same window
-/// and put a later-closing bar in an earlier slot.
-///
-/// One owner for the rule, so the two trims above cannot drift apart about
-/// where the seam is.
-fn seam_bucket_ms(
-    first_engine_bar: Option<&quantick_engine::Bar>,
-    partial: Option<&quantick_engine::Bar>,
-    interval_ms: i64,
-) -> Option<i64> {
-    let first = first_engine_bar.or(partial)?;
-    Some(crate::resample::bucket_start(first.open_time, interval_ms))
-}
-
-/// Whether the chart can reach further back for venue candles, and when it
-/// cannot, why — see [`Tab::older_candles`].
-#[derive(Debug, Clone, Copy, PartialEq, Eq)]
-pub enum OlderCandles {
-    /// Another span can be asked for.
-    Available,
-    /// This feed publishes no candle history at all.
-    FeedServesNone,
-    /// Nothing on this chart is cut by time, so no venue candle was ever
-    /// wanted — the prefix follows what a pane *shows*, not which pane it is.
-    NoChartCutByTime,
-    /// A request is out; the answer is what to wait for.
-    Fetching,
-    /// The opening span has not landed yet. There is nothing to reach back
-    /// *from* until it does.
-    NotArrivedYet,
-    /// A reach-back came back complete with nothing older in it. That is the
-    /// venue's record, or the provider's, and it is the one reason here that
-    /// had to be learned by asking.
-    RecordStartsHere,
-}
-
-impl OlderCandles {
-    /// Whether the control is live.
-    #[must_use]
-    pub const fn is_available(self) -> bool {
-        matches!(self, Self::Available)
-    }
-
-    /// What to tell the trader hovering a control this state disabled.
-    /// `None` when it is not disabled.
-    #[must_use]
-    pub const fn why_not(self) -> Option<&'static str> {
-        match self {
-            Self::Available => None,
-            Self::FeedServesNone => Some("this feed publishes no candle history"),
-            Self::NoChartCutByTime => Some(
-                "no chart here is cut by time, so there are no venue candles \
-                 to extend — switch on the venue lead-in to put them in front \
-                 of a chart cut by trades",
-            ),
-            Self::Fetching => Some("a request is already out; this is what it is fetching"),
-            Self::NotArrivedYet => Some("the first span has not arrived yet"),
-            Self::RecordStartsHere => Some("this is as far back as the venue's record goes"),
         }
     }
 }
@@ -715,45 +559,6 @@ pub struct Tab {
     collapsed_rail: Option<egui::Rect>,
 }
 
-/// Which pane order entry belongs to this frame.
-///
-/// Every visible pane is a trading surface (§11): a price level is as true
-/// on a context chart as on the flow chart, so the aim follows the
-/// **pointer** rather than focus, and holding the buy modifier over any
-/// pane places there without a focusing click first. That is the whole
-/// change from "the focused pane trades" — a trader reading a level on the
-/// context chart no longer has to click it into focus before they can act
-/// on what they just saw.
-///
-/// A drag in flight overrides the pointer. The grabbed line is read against
-/// one pane's price scale, and letting the pointer cross into a neighbour
-/// mid-drag would reprice the order to whatever *that* pane's scale says —
-/// a stop that jumps because the hand strayed. So the pane that started the
-/// drag keeps it until the release.
-///
-/// With the pointer outside every pane (over the dock, off the window, or
-/// on the chrome between panes) the focused pane answers, which is also the
-/// unsplit case: one pane, always the answer, nothing changes.
-fn trading_pane(
-    pointer: Option<egui::Pos2>,
-    panes: &[(PaneIndex, egui::Rect)],
-    dragging: bool,
-    pinned: Option<PaneIndex>,
-    focused: PaneIndex,
-) -> PaneIndex {
-    if dragging && let Some(pinned) = pinned {
-        return pinned;
-    }
-    pointer
-        .and_then(|pointer| {
-            panes
-                .iter()
-                .find(|(_, rect)| rect.contains(pointer))
-                .map(|(side, _)| *side)
-        })
-        .unwrap_or(focused)
-}
-
 impl Tab {
     /// Put the context column away, or bring it back.
     ///
@@ -897,7 +702,7 @@ impl Tab {
             connection_since_ms: metrics::wall_clock_ms(),
             resume_floor_ms: None,
             forced_stall: stall::ForcedStall::from_env(),
-            pending_demo_gap_ms: feed::demo_gap_ms(),
+            pending_demo_gap_ms: quantick_feed::demo_gap_ms(),
             feed_gaps: Vec::new(),
             feed_connection: FeedConnectionState::Connecting,
             feed_capabilities: feed.capabilities,
@@ -1100,565 +905,6 @@ impl Tab {
         // the value it was born with, which is what a test wants and what
         // `unsplit_latency` relies on in production.
         drop(tx);
-    }
-
-    /// Whether any pane on this tab cuts bars by a foldable time interval —
-    /// the gate for venue candle history. Capability-shaped, like every other
-    /// gate in the app (audit S1): the prefix belongs to what a pane *shows*
-    /// (`BarSpec::Time` at a whole number of venue candles), never to which
-    /// pane object it is. `bars → time` on the flow pane earns the same span
-    /// the split's time pane gets.
-    fn any_pane_wants_venue_history(&self) -> bool {
-        std::iter::once(&self.flow_pane)
-            .chain(self.time_pane())
-            .any(|pane| match pane.state.spec().time_interval_ms() {
-                // Cut by time: the venue's candles fold into this pane's own
-                // interval, which is what the prefix has always been. An
-                // interval no whole number of minutes fits into still folds to
-                // nothing, so it still wants none — the lead-in does not change
-                // that, because the pane would draw the answer and discard it.
-                Some(interval) => crate::resample::is_foldable(interval),
-                // Cut by trades: no fold exists, so candles are wanted only
-                // when the trader asked for the lead-in that installs them
-                // unfolded.
-                None => self.venue_lead_in,
-            })
-    }
-
-    /// Ask the venue for its candle history, if there is anything to ask.
-    ///
-    /// Gated on the capability, never on the provider: a feed that serves no
-    /// candles is not asked, and neither is a recording with no context file
-    /// beside it — but a recording that *has* one is, because that file is the
-    /// run-up it was downloaded to carry. One request at a time, and a base
-    /// already held is not re-fetched: changing a pane's interval is a
-    /// different fold over the same bars.
-    fn request_ohlcv_history(&mut self, config: &AppConfig) {
-        let progressive = self.progressive_history;
-        // Not gated on the source. A recording answers this from the context
-        // file downloaded beside it — the run-up it exists to carry — and the
-        // capability is already false on one that has none, so a replay
-        // without context is simply never asked. Refusing here instead meant
-        // a recording opened with no context at all and only picked it up if
-        // the trader happened to press *load older*.
-        if !self.any_pane_wants_venue_history()
-            || self.ohlcv_pending
-            || self.ohlcv_base.is_some()
-            || !self.capabilities(config).ohlcv_history
-        {
-            return;
-        }
-        let slice_ms = progressive.then_some(quantick_feed::OHLCV_SLICE_SPAN_MS);
-        let command = FeedCommand::FetchOhlcv {
-            span_ms: quantick_feed::TIME_HISTORY_SPAN_MS,
-            slice_ms,
-            // The opening request: back from the live edge. Reaching further
-            // than one span is `request_older_ohlcv_history`'s job.
-            before_ms: None,
-        };
-        match self.commands.try_send(command) {
-            Ok(()) => {
-                self.ohlcv_pending = true;
-                self.ohlcv_reaching_back = None;
-                self.ohlcv_older_exhausted = false;
-                self.loading.begin(LoadingTask::VenueHistory);
-                tracing::info!(
-                    target: "quantick::app",
-                    schema_version = 1_u8,
-                    event_code = "OHLCV_REQUESTED",
-                    tab = self.id,
-                    symbol = %self.symbol,
-                    span_ms = quantick_feed::TIME_HISTORY_SPAN_MS,
-                    slice_ms = slice_ms.unwrap_or(0),
-                    action = if progressive { "await_slices" } else { "await_single_reply" },
-                    "asked the venue for candle history"
-                );
-            }
-            // A full channel is a busy frame, and the every-frame poll asks
-            // again. A closed one is a feed thread that is gone, which is
-            // worth saying out loud: nothing will answer, ever.
-            Err(mpsc::error::TrySendError::Full(_)) => tracing::debug!(
-                target: "quantick::app",
-                event_code = "OHLCV_REQUEST_BACKPRESSURE",
-                tab = self.id,
-                action = "retry_next_frame",
-                "candle-history request not queued; channel full"
-            ),
-            Err(mpsc::error::TrySendError::Closed(_)) => tracing::warn!(
-                target: "quantick::app",
-                schema_version = 1_u8,
-                event_code = "OHLCV_REQUEST_CHANNEL_CLOSED",
-                tab = self.id,
-                symbol = %self.symbol,
-                action = "no_history_until_feed_restart",
-                "candle-history request cannot be sent; the feed is gone"
-            ),
-        }
-    }
-
-    /// Watch the capability for the false→true edge and ask when it lands.
-    ///
-    /// Called every frame: the check is two bools and an `Option` when there
-    /// is nothing to do.
-    pub fn poll_ohlcv_capability(&mut self, config: &AppConfig) {
-        let capabilities = self.capabilities(config);
-        let capable = capabilities.ohlcv_history;
-        let rising = capable && !self.ohlcv_capable;
-        self.ohlcv_capable = capable;
-        if capabilities.ohlcv_generation != self.ohlcv_generation {
-            // The venue re-answered. A reconnect can carry a longer block than
-            // the one held, or a corrected one, so what is held goes whether or
-            // not it had bars in it — the guard below then lets a fresh request
-            // through, and the reply reinstalls the prefix by the same path the
-            // first one took.
-            self.ohlcv_generation = capabilities.ohlcv_generation;
-            self.ohlcv_base = None;
-            // And with it, everything learned by reaching back through it. The
-            // oldest bucket a request was measured against is gone, so a reply
-            // still in flight must not be compared to it; and "the record
-            // starts here" was a fact about a block this generation replaces.
-            // `attach` clears both for the same reason on a change of market.
-            self.ohlcv_reaching_back = None;
-            self.ohlcv_older_exhausted = false;
-            // Slices of the discarded answer may still be on their way. They
-            // describe a base that no longer exists, so they are dropped
-            // rather than folded onto nothing.
-            self.ohlcv_stale = self.ohlcv_pending;
-        }
-        if rising {
-            // A session that narrowed *into* serving candles may have answered
-            // an earlier request with nothing; that answer described a feed
-            // that did not know itself yet. Only the edge clears it — a
-            // steady-state empty base is a real "the venue has none".
-            if self.ohlcv_base.as_ref().is_some_and(Vec::is_empty) {
-                self.ohlcv_base = None;
-            }
-        }
-        // Unconditional: every guard inside makes this a no-op once the
-        // request is out or answered, and asking here is what actually retries
-        // a request the command channel refused. The feed ignores a duplicate
-        // while one is in flight, and `ohlcv_pending` means we never send one.
-        self.request_ohlcv_history(config);
-    }
-
-    /// Take a candle-history reply, and put it in front of the time pane.
-    ///
-    /// An empty reply is a complete answer — the venue has none, the provider
-    /// serves none, or the fetch failed — and is recorded as such so the tab
-    /// stops asking. Either way the wait ends on the *closing* slice: a
-    /// provider that answered only on success would strand the spinner on the
-    /// one case that most needs explaining, and one that ended it on the first
-    /// of thirteen slices would claim a quarter of history had arrived when a
-    /// week of it had.
-    ///
-    /// Progressive slices run newest-first, so each one goes in *front* of the
-    /// base already held. The prefix is rebuilt after every slice, which is
-    /// the whole point — the chart grows leftwards while the rest is still
-    /// being fetched, and nothing the trader is looking at moves under them
-    /// (`ChartPane::install_history_prefix` shifts the viewport and every
-    /// bar-anchored drawing by the same amount the prefix grew).
-    fn take_ohlcv_history(
-        &mut self,
-        interval_ms: i64,
-        bars: Vec<quantick_engine::Bar>,
-        slice: quantick_feed::OhlcvSlice,
-    ) {
-        let last = slice.is_last();
-        if self.ohlcv_stale {
-            // An answer the tab already threw away (see `poll_ohlcv_capability`).
-            // The closing slice is what re-opens the door to asking again.
-            if last {
-                self.ohlcv_stale = false;
-                self.ohlcv_pending = false;
-                // Whatever this answer was measured against no longer exists.
-                self.ohlcv_reaching_back = None;
-                self.loading.end(LoadingTask::VenueHistory);
-            }
-            tracing::debug!(
-                target: "quantick::app",
-                event_code = "OHLCV_SLICE_DISCARDED",
-                tab = self.id,
-                bars = bars.len(),
-                last,
-                action = "await_fresh_request",
-                "dropped a slice of a candle answer that was superseded"
-            );
-            return;
-        }
-        if !last {
-            self.take_ohlcv_slice(interval_ms, bars);
-            return;
-        }
-        self.ohlcv_pending = false;
-        if slice == quantick_feed::OhlcvSlice::Refused {
-            // Nobody looked, so nothing is known. End the wait — the spinner
-            // was raised before the command left — and touch nothing else: no
-            // short-answer warning about a venue that never answered, no
-            // refold of the whole base over an empty vector, and above all no
-            // verdict on a reach-back that was never served.
-            self.ohlcv_reaching_back = None;
-            self.loading.end(LoadingTask::VenueHistory);
-            tracing::info!(
-                target: "quantick::app",
-                schema_version = 1_u8,
-                event_code = "OHLCV_REFUSED",
-                tab = self.id,
-                symbol = %self.symbol,
-                action = "await_the_running_fetch",
-                "the provider was already fetching; this request was not served"
-            );
-            return;
-        }
-        let complete = matches!(slice, quantick_feed::OhlcvSlice::Last { complete } if complete);
-        if !complete {
-            // Known-short, not merely short: a venue that stopped answering
-            // partway, or a block clipped to a cap. An instrument younger than
-            // the span is a *complete* answer with fewer bars, which is why
-            // this is carried rather than guessed from the count.
-            //
-            // Said in the log today. §11 records where it will be said on
-            // screen — beside the three-way bar count — and the badge itself
-            // is deferred rather than half-built.
-            tracing::warn!(
-                target: "quantick::app",
-                schema_version = 1_u8,
-                event_code = "OHLCV_INCOMPLETE",
-                tab = self.id,
-                symbol = %self.symbol,
-                interval_ms,
-                bars = bars.len(),
-                action = "install_what_arrived",
-                "the venue's candle history stopped short of the span asked for"
-            );
-        }
-        self.loading.end(LoadingTask::VenueHistory);
-        tracing::info!(
-            target: "quantick::app",
-            schema_version = 1_u8,
-            event_code = "OHLCV_RECEIVED",
-            tab = self.id,
-            symbol = %self.symbol,
-            interval_ms,
-            bars = bars.len(),
-            action = if bars.is_empty() { "no_prefix" } else { "install_prefix" },
-            "candle history arrived"
-        );
-        let usable = self.take_ohlcv_slice(interval_ms, bars);
-        // A *load older* that closed: did the oldest candle actually move? A
-        // reply that re-sends what is already held is not an empty reply, so
-        // the bar count cannot answer this and the oldest bucket can. Latched
-        // here rather than asked again, because the only way to find out was
-        // to ask.
-        if let Some(was_oldest) = self.ohlcv_reaching_back.take() {
-            let now_oldest = self.oldest_venue_candle_ms();
-            let moved = now_oldest.is_some_and(|oldest| oldest < was_oldest);
-            // Only a *complete* answer teaches anything about where the record
-            // starts. A run that came up short — a venue that stopped
-            // answering, a socket that failed — brought nothing older for a
-            // reason that has nothing to do with the venue's depth, and
-            // latching on it would retire the button for the session and tell
-            // the trader their history starts here. Short means try again.
-            // Complete *and* usable. A run that came up short brought nothing
-            // older for a reason unrelated to the venue's depth, and a reply
-            // the pane had to refuse is not an answer about depth at all.
-            self.ohlcv_older_exhausted = complete && usable && !moved;
-            tracing::info!(
-                target: "quantick::app",
-                schema_version = 1_u8,
-                event_code = "OHLCV_OLDER_SETTLED",
-                tab = self.id,
-                symbol = %self.symbol,
-                was_oldest_ms = was_oldest,
-                now_oldest_ms = now_oldest.unwrap_or(0),
-                complete,
-                usable,
-                moved,
-                exhausted = self.ohlcv_older_exhausted,
-                action = if self.ohlcv_older_exhausted {
-                    "stop_offering_older"
-                } else if moved {
-                    "older_available"
-                } else {
-                    "no_evidence_try_again"
-                },
-                "a request for older candles settled"
-            );
-        }
-    }
-
-    /// Merge one slice into the base and rebuild the prefix from it.
-    ///
-    /// Reports whether the slice was *usable* — an interval this pane can fold
-    /// from. A refused slice is not an answer about the venue's depth, and the
-    /// exhaustion latch upstream must not read it as one.
-    ///
-    /// Shared by the closing reply and every slice before it: whether an
-    /// answer arrived whole or in thirteen pieces changes when the wait ends,
-    /// never how the candles are installed. Merging rather than replacing is
-    /// what makes that true — a run's last slice is the *oldest* week of the
-    /// span, and assigning it over the base would throw away the twelve that
-    /// had already been drawn.
-    fn take_ohlcv_slice(&mut self, interval_ms: i64, bars: Vec<quantick_engine::Bar>) -> bool {
-        if interval_ms != quantick_feed::OHLCV_BASE_INTERVAL_MS && !bars.is_empty() {
-            // The event tags its own interval so a consumer never has to
-            // guess; a base this fold was not written for is refused rather
-            // than folded wrongly.
-            tracing::warn!(
-                target: "quantick::app",
-                schema_version = 1_u8,
-                event_code = "OHLCV_UNEXPECTED_BASE",
-                interval_ms,
-                expected_ms = quantick_feed::OHLCV_BASE_INTERVAL_MS,
-                action = if self.ohlcv_base.is_some() {
-                    "refuse_slice_keep_prefix"
-                } else {
-                    "no_prefix"
-                },
-                "candle history arrived at an interval the pane cannot fold from"
-            );
-            // Refuse the slice, never the history. Recording an empty base is
-            // right for the *opening* answer — it is how "this venue serves
-            // nothing this pane can fold" is remembered — but a reach-back
-            // reply at a bad interval arrives on top of a week (or a quarter)
-            // the trader has already paged in, and throwing that away would
-            // cost them everything they waited for over one unusable slice.
-            if self.ohlcv_base.is_none() {
-                self.ohlcv_base = Some(Vec::new());
-            }
-            return false;
-        }
-        let base = self.ohlcv_base.get_or_insert_with(Vec::new);
-        if base.is_empty() {
-            *base = bars;
-        } else if !bars.is_empty() {
-            merge_older_candles(base, bars);
-        }
-        self.refold_history_prefix();
-        true
-    }
-
-    /// Hand the tab one candle-history reply directly, as the feed drain does.
-    ///
-    /// The `QUANTICK_VENUE_HISTORY_DEMO` hook's one door in. It goes through
-    /// the same function a real reply does — a hook that installed a prefix by
-    /// another route would photograph a state the app cannot actually be in.
-    /// The pending flags are set first because that is what a request having
-    /// gone out looks like, and an unclosed run is precisely the frame the
-    /// `partial` variant exists to reach.
-    pub fn deliver_ohlcv_slice(
-        &mut self,
-        interval_ms: i64,
-        bars: Vec<quantick_engine::Bar>,
-        slice: quantick_feed::OhlcvSlice,
-    ) {
-        if !self.ohlcv_pending {
-            self.ohlcv_pending = true;
-            self.loading.begin(LoadingTask::VenueHistory);
-        }
-        self.take_ohlcv_history(interval_ms, bars, slice);
-    }
-
-    /// How many venue candles this tab holds, at the base interval. Zero on a
-    /// feed that serves none, and before the first reply lands.
-    #[must_use]
-    pub fn venue_candles_held(&self) -> usize {
-        self.ohlcv_base.as_ref().map_or(0, Vec::len)
-    }
-
-    /// The oldest venue candle held, by bucket start.
-    fn oldest_venue_candle_ms(&self) -> Option<i64> {
-        self.ohlcv_base.as_ref()?.first().map(|bar| bar.open_time)
-    }
-
-    /// Whether asking for older candles could get the trader anything — and
-    /// when it could not, *which* of the reasons it is.
-    ///
-    /// A bool would be enough to grey the control out and is not enough to
-    /// say why, and "why" is the whole of the disabled tooltip. Under the
-    /// data-honesty rule a control that offers three possible reasons and is
-    /// disabled for a fourth is telling the trader something untrue: on a tick
-    /// chart the answer is not "the venue's record starts here", it is that
-    /// nothing on this chart is cut by time, so no venue candle was ever
-    /// wanted. One enum, so the reason shown is the reason.
-    #[must_use]
-    pub fn older_candles(&self, capabilities: FeedCapabilities) -> OlderCandles {
-        if !capabilities.ohlcv_history {
-            OlderCandles::FeedServesNone
-        } else if !self.any_pane_wants_venue_history() {
-            OlderCandles::NoChartCutByTime
-        } else if self.ohlcv_pending {
-            OlderCandles::Fetching
-        } else if self.oldest_venue_candle_ms().is_none() {
-            OlderCandles::NotArrivedYet
-        } else if self.ohlcv_older_exhausted {
-            OlderCandles::RecordStartsHere
-        } else {
-            OlderCandles::Available
-        }
-    }
-
-    /// Put the venue's candles in front of a chart cut by trades, or take
-    /// them away again.
-    ///
-    /// The named call behind the window's switch: it refolds on the spot, so
-    /// the answer is on screen in the frame that flips it rather than at the
-    /// next candle arrival. Idempotent, and cheap when nothing changes.
-    pub fn set_venue_lead_in(&mut self, enabled: bool) {
-        if self.venue_lead_in == enabled {
-            return;
-        }
-        self.venue_lead_in = enabled;
-        self.refold_history_prefix();
-    }
-
-    /// Shorthand for the one caller that only needs the yes/no.
-    #[must_use]
-    pub fn can_load_older_candles(&self, capabilities: FeedCapabilities) -> bool {
-        self.older_candles(capabilities).is_available()
-    }
-
-    /// Reach one more [`quantick_feed::TIME_HISTORY_SPAN_MS`] into the past and
-    /// prepend what comes back.
-    ///
-    /// The whole reason a chart opens on a week rather than a quarter: the
-    /// deep history is available, it is simply asked for by the trader who
-    /// wants it. Each call is the same request the opening one was, with its
-    /// right-hand edge moved to just before the oldest candle held — so the
-    /// windows never overlap and the merge on the way back in has nothing to
-    /// reconcile.
-    ///
-    /// Reports whether a request actually went out, so a caller can tell "the
-    /// venue is fetching" from "there was nothing to ask for".
-    pub fn request_older_ohlcv_history(&mut self, capabilities: FeedCapabilities) -> bool {
-        let oldest = self.oldest_venue_candle_ms();
-        if !self.can_load_older_candles(capabilities) || oldest.is_none() {
-            tracing::debug!(
-                target: "quantick::app",
-                event_code = "OHLCV_OLDER_DECLINED",
-                tab = self.id,
-                pending = self.ohlcv_pending,
-                exhausted = self.ohlcv_older_exhausted,
-                held = oldest.is_some(),
-                "nothing older to ask for"
-            );
-            return false;
-        }
-        // Read once above and unwrapped here: `can_load_older_candles` already
-        // required it, and a second read that could disagree with the guard is
-        // exactly the kind of duplicate this branch is trying not to leave.
-        let oldest = oldest.expect("the guard above required a candle held");
-        // One millisecond before the oldest bucket start held. The plan's
-        // windows are closed at both ends, so anything else would re-fetch the
-        // candle already on screen.
-        let before_ms = oldest.saturating_sub(1);
-        let slice_ms = self
-            .progressive_history
-            .then_some(quantick_feed::OHLCV_SLICE_SPAN_MS);
-        let command = FeedCommand::FetchOhlcv {
-            span_ms: quantick_feed::TIME_HISTORY_SPAN_MS,
-            slice_ms,
-            before_ms: Some(before_ms),
-        };
-        match self.commands.try_send(command) {
-            Ok(()) => {
-                self.ohlcv_pending = true;
-                self.ohlcv_reaching_back = Some(oldest);
-                self.loading.begin(LoadingTask::VenueHistory);
-                tracing::info!(
-                    target: "quantick::app",
-                    schema_version = 1_u8,
-                    event_code = "OHLCV_OLDER_REQUESTED",
-                    tab = self.id,
-                    symbol = %self.symbol,
-                    span_ms = quantick_feed::TIME_HISTORY_SPAN_MS,
-                    before_ms,
-                    slice_ms = slice_ms.unwrap_or(0),
-                    action = "await_prepend",
-                    "asked the venue for another span of older candles"
-                );
-                true
-            }
-            // Same two cases the opening request distinguishes, and for the
-            // same reasons: a busy frame is worth a retry, a dead feed is
-            // worth saying out loud.
-            Err(mpsc::error::TrySendError::Full(_)) => {
-                tracing::debug!(
-                    target: "quantick::app",
-                    event_code = "OHLCV_OLDER_BACKPRESSURE",
-                    tab = self.id,
-                    action = "retry_on_next_click",
-                    "older-candle request not queued; channel full"
-                );
-                false
-            }
-            Err(mpsc::error::TrySendError::Closed(_)) => {
-                tracing::warn!(
-                    target: "quantick::app",
-                    schema_version = 1_u8,
-                    event_code = "OHLCV_OLDER_CHANNEL_CLOSED",
-                    tab = self.id,
-                    symbol = %self.symbol,
-                    action = "no_history_until_feed_restart",
-                    "older-candle request cannot be sent; the feed is gone"
-                );
-                false
-            }
-        }
-    }
-
-    /// Rebuild every pane's prefix from the base at that pane's interval.
-    ///
-    /// Free of the venue: a chip click lands here, not on the network. One
-    /// base, one fold per time-cutting pane — the flow pane showing time
-    /// bars folds exactly as the split's time pane does (audit S1).
-    /// Reports whether any prefix actually changed — an installed prefix
-    /// rebuilds the indicators, so the caller can skip sending a second one.
-    pub fn refold_history_prefix(&mut self) -> bool {
-        let Self {
-            ohlcv_base,
-            flow_pane,
-            time_panes,
-            venue_lead_in,
-            ..
-        } = self;
-        let venue_lead_in = *venue_lead_in;
-        let Some(base) = ohlcv_base.as_ref() else {
-            return false;
-        };
-        let mut changed = false;
-        for pane in std::iter::once(flow_pane).chain(time_panes.iter_mut()) {
-            // A pane not cutting by time has no interval to fold to, and a
-            // sub-minute one has no whole number of venue candles in it: both
-            // get no prefix, which is the honest answer rather than an
-            // invented one.
-            let prefix = match pane.state.spec().time_interval_ms() {
-                Some(interval) => {
-                    let folded = crate::resample::fold(base, interval);
-                    trim_to_seam(
-                        folded,
-                        pane.state.bars().first(),
-                        pane.state.partial(),
-                        interval,
-                    )
-                }
-                // A pane not cutting by time has no interval to fold *to*, so
-                // the base goes in unfolded: the venue's own minutes, in front
-                // of bars cut by trades. Real candles, counted apart from
-                // built bars on the status bar — a minute never becomes a tick
-                // bar, the two sit side by side and each says what it is.
-                // Asked for, never assumed: without the switch the honest
-                // answer is still no prefix at all.
-                None if venue_lead_in => trim_borrowed_to_seam(
-                    base,
-                    pane.state.bars().first(),
-                    pane.state.partial(),
-                    quantick_feed::OHLCV_BASE_INTERVAL_MS,
-                ),
-                None => Vec::new(),
-            };
-            changed |= pane.install_history_prefix(prefix);
-        }
-        changed
     }
 
     /// Whether this tab has trouble worth showing on its chip while it sits in
@@ -2185,610 +1431,6 @@ impl Tab {
         self.forced_latency.or(*self.feed_latency.borrow())
     }
 
-    /// Reach into the past, as far as [`Self::history_reach`] says.
-    ///
-    /// The named call behind the `+ older` button, the overflow entry and the
-    /// `QUANTICK_LOAD_OLDER` hook — one path, so an operator without a mouse
-    /// reaches exactly what a click reaches. Non-blocking: with a reach of one
-    /// page this is the single request it always was, and with a longer reach
-    /// it is the first of a run each reply continues
-    /// ([`Self::settle_history_page`]).
-    pub fn request_older_history(&mut self, config: &AppConfig) {
-        if self.campaign.is_some() {
-            // A run already has its one permitted request out, and the reply
-            // is what sends the next. Pressing again would raise a second wait
-            // on the same indicator and ask the transport for two pages it
-            // will not serve at once.
-            tracing::debug!(
-                target: "quantick::app",
-                event_code = "HISTORY_REACH_ALREADY_RUNNING",
-                tab = self.id,
-                action = "ignore_press",
-                "a reach is already paging; this press changes nothing"
-            );
-            return;
-        }
-        // Whatever the last press had to say is spent the moment this one is
-        // made: the outcome on screen must be the outcome of the press the
-        // trader is waiting on, never the one before it.
-        self.history_note = None;
-        // Read before the request goes out: the anchor is where the chart
-        // reached *before* this run, and everything older arrived because of
-        // it. That is what makes a second press fetch the session before the
-        // one the first press brought in, rather than finding its work done.
-        let anchor_ms = self.oldest_retained_trade_ms();
-        if !self.send_load_older() {
-            // Nothing was even asked, so nothing will answer. Said on screen
-            // rather than only in the log: to the trader this is a press that
-            // did nothing, which is the whole bug.
-            self.raise_history_note(REQUEST_REFUSED_NOTICE);
-            return;
-        }
-        if self.history_reach.runs_a_campaign() {
-            // A chart holding no prints has nothing to page back *from*, so
-            // the single request above is the whole of this press: the next
-            // one, with a tape under it, starts the run.
-            let held = self.flow_pane.state.trades().len();
-            // The trader's live choice outranks the config seed: the toolbar
-            // and the control plane both write the window's value, and a run
-            // started after that must reach what they asked for rather than
-            // what the file said at startup.
-            let bounds = history_reach::ReachBounds {
-                span_ms: i64::from(self.history_reach_span_minutes) * 60_000,
-                ..config.history.reach_bounds()
-            };
-            self.campaign =
-                anchor_ms.map(|anchor| Campaign::new(anchor, held, bounds, self.history_reach));
-        }
-    }
-
-    /// Queue one `load_older` and raise the wait its reply will resolve.
-    ///
-    /// Returns whether the command went out. A refusal is the end of whatever
-    /// asked for it: nothing will answer, so nothing may keep waiting.
-    fn send_load_older(&mut self) -> bool {
-        match self.commands.try_send(FeedCommand::LoadOlder {
-            count: self.history_step.max(1),
-        }) {
-            Ok(()) => {
-                self.loading.begin(LoadingTask::History);
-                tracing::info!(
-                    target: "quantick::app",
-                    count = self.history_step,
-                    reach = self.history_reach.token(),
-                    "requested older history"
-                );
-                true
-            }
-            Err(mpsc::error::TrySendError::Full(_)) => {
-                tracing::debug!(target: "quantick::app", "older-history request already pending");
-                false
-            }
-            Err(mpsc::error::TrySendError::Closed(_)) => {
-                tracing::warn!(target: "quantick::app", "feed command channel closed");
-                false
-            }
-        }
-    }
-
-    /// Slices of the opening session still to arrive, or `None` when the chart
-    /// is not being filled in behind.
-    ///
-    /// Read by the control plane so an operator without a mouse can tell a
-    /// chart that is still arriving from one that has everything it is going
-    /// to get — the same question the trader answers by watching the bars
-    /// grow leftward.
-    #[must_use]
-    pub const fn opening_slices_remaining(&self) -> Option<u64> {
-        self.opening_slices_remaining
-    }
-
-    /// The oldest print this tab still holds.
-    ///
-    /// Read off the flow pane: every pane is fed the same tape and cuts it its
-    /// own way, and the flow pane is the one that always exists.
-    fn oldest_retained_trade_ms(&self) -> Option<i64> {
-        self.flow_pane
-            .state
-            .trades()
-            .first()
-            .map(|trade| trade.timestamp_ms)
-    }
-
-    /// Decide what a page that just landed means for the reach that asked —
-    /// and leave the trader something to read when it means nothing arrived.
-    ///
-    /// Rate: **rare** — once per history reply. The scan inside
-    /// [`Campaign::advance`] stops at the anchor, so its cost is the page that
-    /// arrived rather than the whole retained tape.
-    fn settle_history_page(&mut self, page_len: usize) {
-        let Some(mut campaign) = self.campaign.take() else {
-            if page_len == 0 {
-                self.raise_history_note(self.empty_page_verdict());
-            }
-            return;
-        };
-        // Putting the reach back to one page is how a run is called off. It is
-        // the only stop a trader has — pressing again mid-run is refused, so
-        // the button cannot be a cancel without a double-click becoming one —
-        // and it is where they would look, because it is the control that
-        // started this.
-        if !self.history_reach.runs_a_campaign() {
-            tracing::info!(
-                target: "quantick::app",
-                schema_version = 1_u8,
-                event_code = "HISTORY_REACH_SETTLED",
-                tab = self.id,
-                symbol = %self.symbol,
-                pages = campaign.pages_spent(),
-                anchor_ms = campaign.anchor_ms(),
-                reached_ms = self.oldest_retained_trade_ms().unwrap_or(0),
-                action = "reach_withdrawn",
-                "the reach was put back to one page; the run stops here"
-            );
-            return;
-        }
-        // The feed's own answer, not the configured one: `history_paging` goes
-        // false the moment a venue reports its record exhausted, and asking
-        // again after that spins a run against a wall.
-        let can_page = self.feed_capabilities.borrow().history_paging;
-        match campaign.advance(self.flow_pane.state.trades(), can_page) {
-            CampaignStep::Ask => {
-                if self.send_load_older() {
-                    self.campaign = Some(campaign);
-                } else {
-                    // Nothing will answer, so the run is over. Said out loud
-                    // rather than retried in silence: a closed channel is a
-                    // feed that is gone, and a full one is a frame so busy
-                    // that pressing again is the honest recovery.
-                    tracing::warn!(
-                        target: "quantick::app",
-                        schema_version = 1_u8,
-                        event_code = "HISTORY_REACH_STALLED",
-                        tab = self.id,
-                        symbol = %self.symbol,
-                        pages = campaign.pages_spent(),
-                        action = "stop_and_wait_for_another_press",
-                        "a load-older reach could not queue its next page"
-                    );
-                    self.raise_history_note(REQUEST_REFUSED_NOTICE);
-                }
-            }
-            CampaignStep::Stop(end) => {
-                tracing::info!(
-                    target: "quantick::app",
-                    schema_version = 1_u8,
-                    event_code = "HISTORY_REACH_SETTLED",
-                    tab = self.id,
-                    symbol = %self.symbol,
-                    pages = campaign.pages_spent(),
-                    anchor_ms = campaign.anchor_ms(),
-                    reached_ms = self.oldest_retained_trade_ms().unwrap_or(0),
-                    action = end.action(),
-                    "a load-older reach finished"
-                );
-                // The same verdict the log just took, in the trader's words.
-                // `ReachMet` has none: yesterday is on the chart, and the
-                // chart says it better than a sentence about the chart.
-                if let Some(notice) = end.notice() {
-                    self.raise_history_note(notice);
-                }
-            }
-        }
-    }
-
-    /// Why one page came back empty, in the words the run would have used.
-    ///
-    /// No campaign was behind this reply — the one-page reach, or a longer one
-    /// that had no tape to page back from — so nothing else is going to judge
-    /// it. It reads the same two facts [`Campaign::advance`] reads first, in
-    /// the same order, so the two paths cannot end up telling the trader
-    /// different stories about one empty block:
-    ///
-    /// - the feed having withdrawn paging is [`CampaignEnd::Exhausted`], not a
-    ///   request that happened to come back empty. Saying otherwise would
-    ///   leave a note about this request beside a button that has just greyed
-    ///   itself out, and the trader reading two accounts of one press.
-    /// - a chart with nothing on it is [`CampaignEnd::NothingCharted`]. The
-    ///   venue is not at fault for a press that had no anchor to reach back
-    ///   from, and blaming it there is the data-honesty rule broken in the
-    ///   trader's own words.
-    ///
-    /// Only what is left over is this reply's own: one empty answer, which is
-    /// not evidence that a record is spent.
-    fn empty_page_verdict(&self) -> &'static str {
-        if !self.feed_capabilities.borrow().history_paging {
-            return CampaignEnd::Exhausted.notice().unwrap_or(EMPTY_PAGE_NOTICE);
-        }
-        if self.flow_pane.state.trades().is_empty() {
-            return CampaignEnd::NothingCharted
-                .notice()
-                .unwrap_or(EMPTY_PAGE_NOTICE);
-        }
-        EMPTY_PAGE_NOTICE
-    }
-
-    /// Drop the run this tab was making and the verdict it produced.
-    ///
-    /// One call, because the two always travel together and always mean the
-    /// same thing: the run and everything it had to say belong to a tape this
-    /// tab no longer shows. Split across the three reset paths they were two
-    /// fields a fourth path could half-remember, and the half it forgot would
-    /// hang a sentence about the previous symbol's press over the new one's
-    /// chart.
-    fn abandon_history_run(&mut self) {
-        self.campaign = None;
-        self.history_note = None;
-    }
-
-    /// Put one sentence about the last press where the trader is looking.
-    ///
-    /// `pub(crate)` for the `QUANTICK_HISTORY_NOTE` hook, which photographs
-    /// this surface by raising a real ending's real sentence through this very
-    /// call — the state is otherwise reachable only by pressing the button
-    /// against a venue that happens to be refusing.
-    pub(crate) fn raise_history_note(&mut self, text: &'static str) {
-        self.history_note = Some(HistoryNote {
-            text,
-            raised_at: std::time::Instant::now(),
-        });
-    }
-
-    /// What the last *load older* press had to say, while it is still on
-    /// screen. `None` is the ordinary state: no press yet, a press that landed
-    /// what it promised, or one whose remark has had its time.
-    #[must_use]
-    pub fn history_note(&self) -> Option<&'static str> {
-        self.history_note.map(|note| note.text)
-    }
-
-    /// Drop the note once it has had its [`HISTORY_NOTE_LINGER`].
-    ///
-    /// Rate: **per-frame**, and deliberately trivial — a `Copy` `Option` and
-    /// one duration comparison, both skipped entirely while there is no note.
-    /// Told the time rather than reading a clock, the way `replay` is, so a
-    /// test can walk past the linger without sleeping through it.
-    pub fn expire_history_note(&mut self, now: std::time::Instant) {
-        if self.history_note.is_some_and(|note| {
-            now.saturating_duration_since(note.raised_at) >= HISTORY_NOTE_LINGER
-        }) {
-            self.history_note = None;
-        }
-    }
-
-    /// Whether a run of *load older* requests is in flight.
-    ///
-    /// Read by the toolbar, so the button can say what it is doing rather than
-    /// look idle while pages land behind it.
-    #[must_use]
-    pub const fn history_reach_running(&self) -> bool {
-        self.campaign.is_some()
-    }
-
-    /// Allocate a capture generation well above all reconnect generations from
-    /// the previous UI capture epoch.
-    pub fn next_book_generation(&mut self) -> u64 {
-        self.book_capture_epoch = self.book_capture_epoch.saturating_add(1);
-        self.book_capture_epoch
-            .saturating_mul(BOOK_GENERATION_STRIDE)
-    }
-
-    /// Keep the recorder running for any feed that can stream depth.
-    ///
-    /// Capture is a data concern: it starts with the feed and stops only when
-    /// the market itself changes (feed/symbol switch, or a replay taking the
-    /// chart over). Showing and hiding the map never reaches this far, which is
-    /// what lets a hidden heatmap come back with its history intact.
-    ///
-    /// Recording with nobody watching stays inside the retention budget the
-    /// heatmap already had — `retention_ms` (30 min by default) bounded by
-    /// `max_history_runs` / `max_history_bytes` — so the ceiling is the same
-    /// one an open map pays for, not a new one.
-    ///
-    /// Idempotent and cheap, so the frame loop can call it as a heartbeat on
-    /// top of the lifecycle calls: already recording costs one bool read, and
-    /// a replay costs one more `Option` check.
-    pub fn ensure_book_capture(&mut self, config: &AppConfig) {
-        if self.tape().enabled() || !self.capabilities(config).book_capture {
-            return;
-        }
-        self.request_book_capture(config, true);
-    }
-
-    /// Start or stop the independent depth pipeline without touching aggTrades
-    /// or candle construction. UI state changes only if the command is queued.
-    pub fn request_book_capture(&mut self, config: &AppConfig, enabled: bool) {
-        if !self.capabilities(config).book_capture {
-            tracing::warn!(
-                target: "quantick::app",
-                schema_version = 1_u8,
-                event_code = "HEATMAP_PROVIDER_UNSUPPORTED",
-                feed = self.feed_id.as_str(),
-                symbol = self.symbol.as_str(),
-                enabled,
-                action = "leave_capture_disabled",
-                "selected provider has no order-book pipeline"
-            );
-            return;
-        }
-
-        let generation = self.next_book_generation();
-        let command = FeedCommand::SetBookCapture {
-            enabled,
-            initial_generation: generation,
-        };
-        match self.commands.try_send(command) {
-            Ok(()) => self.tape_mut().set_enabled(enabled, generation),
-            Err(mpsc::error::TrySendError::Full(_)) => tracing::warn!(
-                target: "quantick::app",
-                schema_version = 1_u8,
-                event_code = "HEATMAP_COMMAND_BACKPRESSURE",
-                symbol = self.symbol.as_str(),
-                enabled,
-                generation,
-                action = "retry_on_next_frame",
-                "book capture command channel is full"
-            ),
-            Err(mpsc::error::TrySendError::Closed(_)) => tracing::warn!(
-                target: "quantick::app",
-                schema_version = 1_u8,
-                event_code = "HEATMAP_COMMAND_CHANNEL_CLOSED",
-                symbol = self.symbol.as_str(),
-                enabled,
-                generation,
-                action = "keep_current_capture_state",
-                "book capture command channel is closed"
-            ),
-        }
-    }
-
-    /// Restart capture after a semantic configuration change such as base
-    /// price grouping. The view commits its staged reset only after this
-    /// command is accepted, preserving current history on backpressure.
-    pub fn restart_book_capture(&mut self) {
-        if !self.tape().enabled() {
-            return;
-        }
-        let generation = self.next_book_generation();
-        match self.commands.try_send(FeedCommand::RestartBookCapture {
-            initial_generation: generation,
-        }) {
-            Ok(()) => self.tape_mut().accept_capture_grouping_restart(generation),
-            Err(mpsc::error::TrySendError::Full(_)) => {
-                self.tape_mut()
-                    .reject_capture_grouping_restart("command_channel_full");
-                tracing::warn!(
-                    target: "quantick::app",
-                    schema_version = 1_u8,
-                    event_code = "HEATMAP_RESTART_BACKPRESSURE",
-                    symbol = self.symbol.as_str(),
-                    generation,
-                    action = "keep_existing_capture",
-                    "book restart command channel is full"
-                );
-            }
-            Err(mpsc::error::TrySendError::Closed(_)) => {
-                self.tape_mut()
-                    .reject_capture_grouping_restart("command_channel_closed");
-                tracing::warn!(
-                    target: "quantick::app",
-                    schema_version = 1_u8,
-                    event_code = "HEATMAP_RESTART_CHANNEL_CLOSED",
-                    symbol = self.symbol.as_str(),
-                    generation,
-                    action = "keep_existing_capture",
-                    "book restart command channel is closed"
-                );
-            }
-        }
-    }
-
-    /// Respawn the feed and reset the chart when the selected feed or symbol
-    /// differs from what is currently streaming. A no-op otherwise.
-    pub fn maybe_switch_feed(&mut self, config: &AppConfig) {
-        // A replay owns the chart until it is closed. The selectors are not
-        // drawn while it plays, so nothing can diverge here — but a stale
-        // selection must not respawn a live feed underneath the recording.
-        if self.replay.is_some() {
-            return;
-        }
-        if self.active == (self.feed_id.clone(), self.symbol.clone()) {
-            return;
-        }
-        let (previous_feed, previous_symbol) = self.active.clone();
-        let Some(provider) = config.provider_of(&self.feed_id) else {
-            tracing::warn!(
-                target: "quantick::app",
-                feed = %self.feed_id,
-                "selected feed is not in the config; ignoring switch"
-            );
-            // Snap the selection back to what is actually running.
-            (self.feed_id, self.symbol) = self.active.clone();
-            return;
-        };
-        // The recorder follows the feed, not the toggle: a market that can
-        // stream depth is recorded from the moment it starts streaming. Kept
-        // for the log line below; the start itself goes through
-        // [`Self::ensure_book_capture`], the one place that decides it.
-        let resume_book_capture = provider.capabilities().book_capture;
-
-        tracing::info!(
-            target: "quantick::app",
-            schema_version = 1_u8,
-            event_code = "FEED_SWITCH",
-            feed = %self.feed_id,
-            symbol = %self.symbol,
-            provider = ?provider,
-            resume_book_capture,
-            action = "reset_market_state",
-            "switching feed/symbol; resetting chart"
-        );
-
-        // Dropping the old handle stops the old feed thread. The new feed starts
-        // with a fresh backfill in flight.
-        let handle = feed::spawn_live(provider, &self.symbol, &config.metatrader, shelf_dir());
-        self.attach(handle);
-
-        // Rebuild every pane from scratch for the new stream, each keeping its
-        // own bar spec. Retained trades from the old symbol must not leak in.
-        for pane in self.panes_mut() {
-            pane.reset_series();
-        }
-        self.drop_overlay_gestures();
-        // The marks stay — only the trader deletes a drawing — but they were
-        // drawn on the instrument that just left. Time survives a symbol
-        // switch and price does not, so without this a BTC level would paint
-        // at full strength over an index chart, at a number that means
-        // nothing there (§D7b).
-        for pane in self.panes_mut() {
-            pane.drawings.mark_market_changed();
-            // The regions those instances watched belong to the market that
-            // just left; a bot must never fire on a level from another
-            // instrument. Disarmed by name, never silently dropped. Their
-            // pending entries need no sweep of ours: the paper reset a few
-            // lines down cancels every order with the honest `reset` label.
-            let _ = pane
-                .strategies
-                .disarm_all(quantick_strategy::DisarmReason::MarketChanged);
-            let _ = pane.take_strategy_bars();
-        }
-        self.history_trades = 0;
-        // The old feed's unanswered loads died with its channel; the new feed
-        // opens with exactly one backfill in flight.
-        self.opening_slices_remaining = None;
-        self.loading.restart(LoadingTask::History);
-        self.latest_trade_latency_ms = None;
-        let symbol = self.symbol.clone();
-        self.tape_mut().reset_for_symbol(symbol);
-        // The simulated position cannot follow this tab onto another market:
-        // flatten at the old symbol's last mark, labeled and journaled (the
-        // journal still targets the old symbol — it only follows `self.symbol`
-        // at the top of the next drain).
-        self.paper.on_timeline_reset();
-
-        self.active = (self.feed_id.clone(), self.symbol.clone());
-        self.refresh_chip_label(config);
-        self.ensure_book_capture(config);
-        self.apply_feed_bubble_preset_after_switch(config, &previous_feed, &previous_symbol);
-    }
-
-    /// Apply the arrived-at declared preset — when the switch crossed feeds,
-    /// or when it crossed symbols whose declared looks differ. A symbol hop
-    /// between two symbols that declare nothing of their own keeps the user's
-    /// panel tweaks, exactly as before per-symbol declarations existed: the
-    /// declared look belongs to the feed, and to the symbols that state one.
-    ///
-    /// One asymmetry is deliberate: hopping *off* a declared symbol onto one
-    /// that declares nothing (on a feed that also declares nothing) keeps the
-    /// look just left on screen — nothing remembers what the panel wore
-    /// before the declaration applied, and inventing a "previous look" store
-    /// for that one hop would be a second owner for the panel's state.
-    pub fn apply_feed_bubble_preset_after_switch(
-        &mut self,
-        config: &AppConfig,
-        previous_feed: &str,
-        previous_symbol: &str,
-    ) {
-        if previous_feed == self.feed_id {
-            let feed = config.feed(&self.feed_id);
-            let arrived = feed.and_then(|feed| feed.bubble_preset_for(&self.symbol));
-            let left = feed.and_then(|feed| feed.bubble_preset_for(previous_symbol));
-            if arrived.is_none() || arrived == left {
-                return;
-            }
-        }
-        self.apply_feed_bubble_preset(config);
-    }
-
-    /// Apply the bubble preset declared for the current feed and symbol, if
-    /// one is declared ([`FeedConfig::bubble_preset_for`]'s ladder: the
-    /// symbol's own entry first, the feed-wide declaration behind it).
-    ///
-    /// A feed declaring nothing changes nothing: the panel keeps the look the
-    /// user last chose. An unknown name is reported and ignored — the presets
-    /// file is user-edited, and a typo there must not silently restyle the
-    /// chart.
-    pub fn apply_feed_bubble_preset(&mut self, config: &AppConfig) {
-        let Some(name) = config
-            .feed(&self.feed_id)
-            .and_then(|feed| feed.bubble_preset_for(&self.symbol))
-            .map(str::to_owned)
-        else {
-            return;
-        };
-        let applied = self.tape_mut().apply_preset(&name);
-        if applied {
-            tracing::info!(
-                target: "quantick::app",
-                schema_version = 1_u8,
-                event_code = "FEED_BUBBLE_PRESET",
-                feed = %self.feed_id,
-                symbol = %self.symbol,
-                preset = name.as_str(),
-                action = "apply_preset",
-                "feed declares a bubble preset; applied"
-            );
-        } else {
-            tracing::warn!(
-                target: "quantick::app",
-                schema_version = 1_u8,
-                event_code = "FEED_BUBBLE_PRESET_UNKNOWN",
-                feed = %self.feed_id,
-                symbol = %self.symbol,
-                preset = name.as_str(),
-                action = "keep_current_look",
-                "feed declares a bubble preset that is not in the presets file; ignoring"
-            );
-        }
-    }
-
-    /// Open wearing the layout the feed declares, if it declares one.
-    ///
-    /// Startup-scoped, like `default_feed`: it decides what a tab on this
-    /// feed *opens* showing, and never touches a layout the user has since
-    /// chosen — the callers are tab creation, nothing else. A feed with no
-    /// `default_layout` changes nothing, so the factory default stays the
-    /// flow pane (the decision on record in the UX audit §3).
-    pub fn apply_feed_declared_layout(&mut self, config: &AppConfig) {
-        let Some(declared) = config
-            .feed(&self.feed_id)
-            .and_then(|feed| feed.default_layout)
-        else {
-            return;
-        };
-        let layout = CanvasLayout::from(declared);
-        // A declared time-bar spec names the interval the declared layout's
-        // time pane opens on; the pane's own header takes over from there.
-        if layout.shows_time()
-            && let Some(BarSpec::Time(ms)) = config.startup_spec_for(&self.feed_id)
-        {
-            self.time_pane_opening_interval_ms = ms;
-        }
-        tracing::info!(
-            target: "quantick::app",
-            schema_version = 1_u8,
-            event_code = "FEED_DECLARED_LAYOUT",
-            feed = %self.feed_id,
-            layout = ?layout,
-            action = "open_declared_layout",
-            "feed declares an opening layout; applied"
-        );
-        self.set_layout(layout);
-        // Opening is not switching. [`Self::set_layout`] focuses whatever the
-        // switch revealed, which is right for a menu click — the trader asked
-        // for that pane. On a tab that has never been on screen there was no
-        // gesture and nothing was revealed, and leaving the focus there put a
-        // fresh window's BARS group and status line on the *context* chart:
-        // the first thing a trader touched changed the timeframe pane instead
-        // of the flow chart quantick exists to show. So the flow pane takes
-        // the focus whenever this layout draws it.
-        self.focus = if layout.shows_flow() {
-            PaneSide::Flow
-        } else {
-            PaneSide::Time(0)
-        };
-    }
-
     /// The layout each pane opens on, from a saved workspace: the flow pane's
     /// now, each context pane's when it is built.
     ///
@@ -3013,2145 +1655,9 @@ impl Tab {
             }
         }
     }
-
-    /// Drain every feed event available this frame into the engine, tracking the
-    /// observed arrival latency and live-trade counts for the metrics.
-    pub fn drain_feed(&mut self) {
-        self.drain_feed_with_clock(metrics::wall_clock_ms);
-    }
-
-    /// Clock-injected drain used to prove that one UI cycle is one observation.
-    pub fn drain_feed_with_clock(&mut self, mut wall_clock_ms: impl FnMut() -> i64) {
-        // The journal follows this tab's symbol; synced before the drain so a
-        // new feed's first trades are never attributed to the old symbol.
-        // Every tab drains every frame, so every journal tracks its own market
-        // whether or not that tab is the one on screen.
-        let Self { paper, symbol, .. } = self;
-        paper.set_symbol(symbol);
-        let mut live = false;
-        let mut received_at_ms = None;
-        loop {
-            match self.events.try_recv() {
-                Ok(FeedEvent::Backfilled(trades)) => {
-                    self.loading.end(LoadingTask::History);
-                    // A session that resumed onto a kept timeline opens by
-                    // replaying its recent window, which the chart already
-                    // holds. Whatever is genuinely newer is forwarded as live,
-                    // in order, exactly as the feed does inside one session.
-                    if self.resume_floor_ms.is_some() {
-                        live |= self.ingest_resumed(&trades);
-                        continue;
-                    }
-                    self.history_trades += trades.len();
-                    // History only seeds the simulator's mark — filling
-                    // against the past would be look-ahead.
-                    if let Some(last) = trades.last() {
-                        self.paper.seed(last);
-                    }
-                    // One tape, every pane: the split multiplies views of the
-                    // market, never the stream behind them.
-                    for pane in self.panes_mut() {
-                        pane.ingest_backfill(&trades);
-                    }
-                }
-                Ok(FeedEvent::HistoryPrepended(trades)) => {
-                    // The reply — even an empty one — answers exactly one
-                    // pending load; the indicator survives until the last one.
-                    self.loading.end(LoadingTask::History);
-                    // The MetaTrader bridge delivers its recovery window on
-                    // this event rather than as backfill, so a resumed session
-                    // is filtered here too. Prepending it would put a block the
-                    // chart already holds in front of the bars it duplicates.
-                    if self.resume_floor_ms.is_some() {
-                        live |= self.ingest_resumed(&trades);
-                        continue;
-                    }
-                    self.history_trades += trades.len();
-                    // Each pane cuts the older trades into its own bars, so
-                    // each shifts its own anchors by its own count.
-                    for pane in self.panes_mut() {
-                        pane.prepend_history(&trades);
-                    }
-                    // The first engine bar just moved backwards in time, and
-                    // the prefix was trimmed against where it used to be. Any
-                    // venue candle now covering a re-cut minute has to go.
-                    self.refold_history_prefix();
-                    // And the reach that asked for this page decides whether
-                    // to ask for another, and what to tell the trader if it
-                    // will not. After the prepend, so it judges the tape the
-                    // trader can actually see.
-                    self.settle_history_page(trades.len());
-                }
-                Ok(FeedEvent::OpeningPrepended { trades, remaining }) => {
-                    // What is left of the fill, so the chart and an operator
-                    // reading the control plane can both say how much of the
-                    // session is still arriving instead of watching a number
-                    // rise with no denominator. Cleared at zero: the field
-                    // means "a fill is running", and a stale count would keep
-                    // saying so after the last slice landed.
-                    self.opening_slices_remaining = match remaining {
-                        Some(0) | None => None,
-                        some => some,
-                    };
-                    // The rest of the opening session, drawn but not counted.
-                    // Everything the reply path does *except* the two things
-                    // that belong to a request: the loading indicator a press
-                    // raised stays up, and the campaign that press started is
-                    // not handed a page it did not fetch.
-                    if self.resume_floor_ms.is_some() {
-                        live |= self.ingest_resumed(&trades);
-                        continue;
-                    }
-                    self.history_trades += trades.len();
-                    for pane in self.panes_mut() {
-                        pane.prepend_history(&trades);
-                    }
-                    self.refold_history_prefix();
-                }
-                Ok(FeedEvent::Live(trade)) => {
-                    if self.resume_floor_ms.is_some() {
-                        live |= self.ingest_resumed(std::slice::from_ref(&trade));
-                        continue;
-                    }
-                    let received_at_ms = *received_at_ms.get_or_insert_with(&mut wall_clock_ms);
-                    self.ingest_live_trade_at(&trade, received_at_ms);
-                    live = true;
-                }
-                Ok(FeedEvent::LiveBatch(trades)) => {
-                    if !trades.is_empty() {
-                        if self.resume_floor_ms.is_some() {
-                            live |= self.ingest_resumed(&trades);
-                            continue;
-                        }
-                        let received_at_ms = *received_at_ms.get_or_insert_with(&mut wall_clock_ms);
-                        for trade in &trades {
-                            self.ingest_live_trade_at(trade, received_at_ms);
-                        }
-                        live = true;
-                    }
-                }
-                Ok(FeedEvent::Reset) => self.reset_market_state(),
-                Ok(FeedEvent::OhlcvHistory {
-                    interval_ms,
-                    bars,
-                    slice,
-                }) => {
-                    self.take_ohlcv_history(interval_ms, bars, slice);
-                }
-                Err(_) => break,
-            }
-        }
-        // One forming-bar update per pane for the whole drain, however many
-        // prints arrived: only its latest value is ever read.
-        if live {
-            for pane in self.panes_mut() {
-                pane.publish_partial();
-            }
-        }
-        // After the bars exist, so the hooked gap has something to sit
-        // between. Costs one `Option` test per drain when the hook is unset,
-        // which is every run but a capture.
-        self.land_demo_gap();
-        // A reset left the marks waiting for bars to anchor to, and this is
-        // the drain that may have just delivered them. One flag test per pane
-        // when nothing is owed.
-        for pane in self.panes_mut() {
-            pane.settle_pending_reanchor();
-        }
-    }
-
-    /// Take the newest feed notice, if the feed sent any this frame.
-    ///
-    /// Level-triggered rather than queued: only the latest state matters, and
-    /// a burst of bridge output must not queue up cards to show one by one.
-    /// A closed channel (a feed with nothing to report) simply yields nothing.
-    pub fn drain_notices(&mut self) {
-        self.drain_notices_at(metrics::wall_clock_ms());
-    }
-
-    /// Clock-injected half, so a test can drive a notice across its stall
-    /// budget without waiting for one.
-    pub fn drain_notices_at(&mut self, now_ms: i64) {
-        while let Ok(notice) = self.notices.try_recv() {
-            let next = match notice {
-                FeedNotice::Connected => {
-                    self.set_connection(FeedConnectionState::Connected, now_ms);
-                    FeedNotice::Clear
-                }
-                FeedNotice::Reconnecting { .. } => {
-                    self.set_connection(FeedConnectionState::Reconnecting, now_ms);
-                    notice
-                }
-                FeedNotice::Working { .. } | FeedNotice::Attention { .. } => notice,
-                FeedNotice::Clear => FeedNotice::Clear,
-            };
-            // Stamped only on a *change*. A supervisor repeating the same line
-            // every few seconds would otherwise keep resetting the very budget
-            // it is supposed to run out of, and the chart would sit on
-            // "connecting" for as long as the retry loop lived — which is the
-            // failure this clock exists to end.
-            if next != self.notice {
-                self.notice = next;
-                self.notice_since_ms = now_ms;
-            }
-        }
-    }
-
-    /// Where a feed's trouble is actually visible: the painted pane with the
-    /// least on it, and how many bar slots that pane holds.
-    ///
-    /// One tab is one feed, so its notice is tab-scoped information — but its
-    /// *consequence* never is. A MetaTrader tab whose time panes are full of
-    /// the terminal's candle history and whose flow pane has not seen one tick
-    /// is starved in exactly one place, and the card explaining it was being
-    /// centred across all three. The trader read a sentence attached to
-    /// nothing, floating over two charts with nothing wrong with them.
-    ///
-    /// Fewest slots wins; the larger pane breaks a tie, so an all-empty canvas
-    /// puts the card on the biggest chart — which is where it landed before
-    /// this existed, and why a single-pane tab looks unchanged.
-    #[must_use]
-    pub fn starved_pane(&self) -> Option<(egui::Rect, usize)> {
-        self.panes()
-            .filter_map(|(pane, _)| pane.last_area.map(|area| (area, pane.slots())))
-            .min_by(|(left_area, left_slots), (right_area, right_slots)| {
-                left_slots
-                    .cmp(right_slots)
-                    .then_with(|| right_area.area().total_cmp(&left_area.area()))
-            })
-    }
-
-    /// Move the transport to `state`, stamping when it got there.
-    ///
-    /// Only a real transition stamps: a provider that reports the same state
-    /// twice has not changed anything, and treating it as a change would reset
-    /// the budget measured from it.
-    fn set_connection(&mut self, state: FeedConnectionState, now_ms: i64) {
-        if self.feed_connection != state {
-            self.feed_connection = state;
-            self.connection_since_ms = now_ms;
-        }
-    }
-
-    /// This tab's own judgement about a feed that has stopped delivering, or
-    /// `None` while it is merely slow.
-    ///
-    /// Deterministic half: the caller supplies wall clock, exactly as
-    /// [`Self::tape_age_at`] does. The decision itself lives in
-    /// [`quantick_feed::stall`] and touches no clock at all.
-    #[must_use]
-    pub fn stall_at(&self, config: &AppConfig, now_ms: i64) -> Option<Stall> {
-        let provider = config.provider_of(&self.feed_id)?;
-        if let Some(forced) = self.forced_stall {
-            return Some(forced.stall(provider, self.feed_display_name(config)));
-        }
-        stall::assess(
-            &StallInput {
-                notice: &self.notice,
-                connection_since_ms: self.connection_since_ms,
-                connection: self.feed_connection,
-                tape_age_ms: self.tape_age_at(now_ms),
-                attached_ms: self.feed_attached_ms,
-                provider,
-                provider_name: self.feed_display_name(config),
-                replaying: self.replay.is_some(),
-            },
-            now_ms,
-        )
-    }
-
-    /// Deterministic half of live ingestion: `received_at_ms` is the UI's epoch
-    /// observation time, supplied explicitly so tests never wait on a clock.
-    ///
-    /// The transport observation is the window's; what the trade does to the
-    /// bars, the tape and the indicators is the pane's.
-    pub fn ingest_live_trade_at(&mut self, trade: &quantick_engine::Trade, received_at_ms: i64) {
-        self.latest_trade_latency_ms =
-            metrics::feed_lag_ms(received_at_ms, Some(trade.timestamp_ms));
-        self.latest_trade_ms = Some(trade.timestamp_ms);
-        self.live_trades += 1;
-        // The simulator taps the same per-trade point the bar engine does, so
-        // paper trading works identically on a live feed and a replay — and on
-        // a tab the user is not looking at, whose position keeps marking
-        // against its own tape.
-        self.paper.on_trade(trade);
-        for pane in self.panes_mut() {
-            pane.ingest_live_trade(trade);
-        }
-        self.run_strategies();
-    }
-
-    /// Evaluate the armed instances against the bars that just closed and
-    /// route the simulator's answers back to them.
-    ///
-    /// Runs inside the same sweep that ingested the trades, so the slots
-    /// queued with each bar and the drawings' anchors are read against one
-    /// cut of the series. Order per pane: the prints' events first (they
-    /// resolve earlier operations), then closed bars oldest-first, each
-    /// instance's commands applied immediately so its own acknowledgement
-    /// reaches it before anything else does. Cost is zero on a chart with
-    /// no instances — the pane queues no bars and the paper host buffers no
-    /// events.
-    fn run_strategies(&mut self) {
-        // The clock is read only when something is listening for it. This
-        // runs once per print and `SystemTime::now` is a syscall: a chart
-        // with no alarm armed — which is every chart that existed before
-        // this feature — must not start paying for one per trade.
-        let now_ms = if self.any_alarm_armed() {
-            metrics::wall_clock_ms()
-        } else {
-            0
-        };
-        self.run_strategies_at(now_ms);
-    }
-
-    /// Whether any instance on this tab carries an alarm. Usually a walk
-    /// over an empty `Vec`.
-    fn any_alarm_armed(&self) -> bool {
-        self.panes().any(|(pane, _side)| {
-            pane.strategies
-                .instances
-                .iter()
-                .any(|instance| instance.alarm.is_some())
-        })
-    }
-
-    /// The sweep with its clock handed in.
-    ///
-    /// Only the alarm's repeat rule reads it — a cooldown is the one thing
-    /// here measured in seconds a human waits rather than in prints — and
-    /// the kernel never reads a clock of its own, so the reading enters
-    /// through this one door and the tests drive it from a fixture.
-    fn run_strategies_at(&mut self, now_ms: i64) {
-        let now_ms = u64::try_from(now_ms).unwrap_or(0);
-        let print_events = self.paper.account_mut().drain_bot_events();
-        let Self {
-            paper,
-            flow_pane,
-            time_panes,
-            ..
-        } = self;
-        let mut watching = 0;
-        let mut sounds: Vec<crate::audio::Cue> = Vec::new();
-        // The alarm's sounds come from `main`; walking every pane in the
-        // context stack rather than a single time pane comes from this branch.
-        // Both are wanted: a strategy armed on the second stacked chart has to
-        // ring like one armed on the first.
-        for pane in std::iter::once(flow_pane).chain(time_panes.iter_mut()) {
-            if pane.strategies.is_empty() {
-                continue;
-            }
-            // The batch to each instance, applying what it answers with —
-            // the self-protection close after a dropped bracket. Applying
-            // a returned command emits no events of its own (kernel
-            // contract), so one echo settles it.
-            if !print_events.is_empty() {
-                for index in 0..pane.strategies.instances.len() {
-                    let responses = pane.strategies.instances[index]
-                        .armed
-                        .on_sim_events(&print_events);
-                    for command in responses {
-                        let events = paper.account_mut().apply_strategy_command(command);
-                        let _ = pane.strategies.instances[index]
-                            .armed
-                            .on_sim_events(&events);
-                    }
-                }
-            }
-            let bars = pane.take_strategy_bars();
-            if !bars.is_empty() {
-                // An instance whose drawing was deleted from a surface that
-                // could not remove it directly dies here, in the sweep.
-                let alive: Vec<crate::drawings::DrawingId> = pane
-                    .strategies
-                    .instances
-                    .iter()
-                    .map(|instance| instance.drawing)
-                    .filter(|id| pane.drawings.index_of(*id).is_some())
-                    .collect();
-                let orphan_cleanup = pane.strategies.drop_orphans(|id| alive.contains(&id));
-                for command in orphan_cleanup {
-                    // A dead drawing's bot must not leave its entry resting
-                    // with no badge over it; swept through the same funnel.
-                    let _ = paper.account_mut().apply_strategy_command(command);
-                }
-            }
-            for (bar, slot) in &bars {
-                for index in 0..pane.strategies.instances.len() {
-                    let drawing = pane.strategies.instances[index].drawing;
-                    // A region that cannot honestly be tested (hidden, off
-                    // its series, another market) holds fire but never
-                    // starves the ruler: the trigger's contract is every
-                    // closed bar, so the gates shut instead of the feed.
-                    let (region, active) = pane.strategy_region(drawing, *slot).unwrap_or((
-                        quantick_strategy::Region::new(
-                            rust_decimal::Decimal::ZERO,
-                            rust_decimal::Decimal::ZERO,
-                        ),
-                        false,
-                    ));
-                    let flat = paper.is_flat();
-                    let commands = pane.strategies.instances[index]
-                        .armed
-                        .on_closed_bar(bar, &region, active, flat);
-                    // The alarm judges the same bar, from the kernel's own
-                    // reading of it rather than from whether an order went
-                    // out — a busy account and a spent one shot silence the
-                    // order, never the signal. Every closed bar is offered,
-                    // qualifying or not: a preview that failed to hold is
-                    // reported here, and this bar's repeat budget resets.
-                    sounds.extend(pane.strategies.instances[index].alarm_on_closed_bar(now_ms));
-                    for command in commands {
-                        let events = paper.account_mut().apply_strategy_command(command);
-                        let _ = pane.strategies.instances[index]
-                            .armed
-                            .on_sim_events(&events);
-                    }
-                }
-            }
-            // The bar still forming, judged for the alarm only. Nothing
-            // here can place an order or move a state machine: the kernel's
-            // preview path is `&self` all the way down, and this is the one
-            // caller of it.
-            //
-            // This whole block is per *print*, so it opens with the question
-            // that costs least: does any instance on this pane even carry an
-            // alarm? A pane full of ordinary strategies answers no and pays
-            // nothing further — not the partial bar's copy, not the
-            // progress read.
-            let any_alarm = pane
-                .strategies
-                .instances
-                .iter()
-                .any(|instance| instance.alarm.is_some());
-            if let Some(partial) = any_alarm.then(|| pane.state.partial().cloned()).flatten() {
-                let progress = pane.state.progress().map(|(progress, _unit)| progress);
-                let slot = pane.closed_slots();
-                for index in 0..pane.strategies.instances.len() {
-                    // Cheapest gate next, before the region is resolved: on
-                    // all but a handful of prints the alarm answers "not
-                    // yet" and this costs one comparison.
-                    let wants = pane.strategies.instances[index]
-                        .alarm
-                        .as_ref()
-                        .is_some_and(|alarm| alarm.wants_forming_check(progress, now_ms));
-                    if !wants {
-                        continue;
-                    }
-                    let drawing = pane.strategies.instances[index].drawing;
-                    let (region, active) = pane.strategy_region(drawing, slot).unwrap_or((
-                        quantick_strategy::Region::new(
-                            rust_decimal::Decimal::ZERO,
-                            rust_decimal::Decimal::ZERO,
-                        ),
-                        false,
-                    ));
-                    sounds.extend(
-                        pane.strategies.instances[index]
-                            .alarm_on_forming_bar(&partial, &region, active, progress, now_ms),
-                    );
-                }
-            }
-            watching += pane.strategies.watching();
-        }
-        paper.account_mut().set_bot_listening(watching > 0);
-        self.pending_alarm_sounds.extend(sounds);
-    }
-
-    /// Apply the cleanup commands the panes' drawing menus queued this
-    /// frame — a disarm or removal over a resting retest limit cancels the
-    /// order. Runs on the UI frame that clicked, not on the next print: a
-    /// cancel that waits for the market to move may lose the race to it.
-    pub fn apply_strategy_cleanup(&mut self) {
-        let Self {
-            paper,
-            flow_pane,
-            time_panes,
-            ..
-        } = self;
-        for pane in std::iter::once(flow_pane).chain(time_panes.iter_mut()) {
-            for command in pane.take_strategy_cleanup() {
-                let _ = paper.account_mut().apply_strategy_command(command);
-            }
-        }
-    }
-
-    /// Throw away everything loaded and wait for the source to refill it.
-    ///
-    /// Sent by a source that rewound — seeking a replay, for instance. The
-    /// chart is rebuilt from the history that follows rather than patched,
-    /// because bars that already closed cannot be reopened.
-    pub fn reset_market_state(&mut self) {
-        for pane in self.panes_mut() {
-            pane.reset_series();
-            // Indicators follow the chart into the empty state; the refill's
-            // Backfilled event replays them (replay seek funnels through here,
-            // so seeking inherits correct indicator behavior for free).
-            pane.send_indicator_rebuild();
-            pane.last_lane_divider_x = None;
-            // Judgements armed on the old timeline do not carry into the
-            // rebuilt one — the same honesty rule the simulator's flatten
-            // follows, with the reason on the badge. No cleanup commands
-            // come back under this reason: `paper.on_timeline_reset` below
-            // sweeps every order with the honest `reset` label.
-            let _ = pane
-                .strategies
-                .disarm_all(quantick_strategy::DisarmReason::TimelineReset);
-        }
-        self.drop_overlay_gestures();
-        // The timeline these describe is the one being rebuilt. A gap
-        // re-anchors itself by market time, so one left behind would paint
-        // itself onto the refilled series as though a reconnect had happened.
-        self.feed_gaps.clear();
-        self.resume_floor_ms = None;
-        self.history_trades = 0;
-        // A run anchored to a tape that no longer exists cannot continue, and
-        // `restart` below drops the waits its outstanding request would have
-        // resolved. Its verdict goes with it.
-        self.abandon_history_run();
-        self.latest_trade_latency_ms = None;
-        self.latest_trade_ms = None;
-        // The refill arrives as one backfill batch; keep the loading indicator
-        // up until it lands. Requests sent to the source before the reset will
-        // never be answered, so the count restarts rather than accumulates.
-        self.opening_slices_remaining = None;
-        self.loading.restart(LoadingTask::History);
-        let symbol = self.symbol.clone();
-        self.tape_mut().reset_for_symbol(symbol);
-        // The simulator flattens at its last mark and says so — a position
-        // cannot honestly survive into a rebuilt timeline.
-        self.paper.on_timeline_reset();
-    }
-
-    /// Everything this tab must settle before it is dropped.
-    ///
-    /// Feeds and workers need nothing: their loops end when the channels go
-    /// with the tab. The simulator does — an open position is state the user
-    /// created, and the honesty contract says it ends in an explicit, labeled,
-    /// journaled flatten, never by silently vanishing with its window.
-    pub fn close(&mut self) {
-        self.paper.on_timeline_reset();
-    }
-
-    /// Drain a bounded number of synchronized depth events. The separate
-    /// channel and budget ensure heatmap work cannot block candle ingestion.
-    pub fn drain_book_feed(&mut self) {
-        self.drain_book_feed_with_clock(metrics::wall_clock_ms);
-    }
-
-    /// Clock-injected depth drain; a burst handled by one UI frame has one
-    /// observation time, matching the trade-side metric and avoiding O(n)
-    /// system-clock reads.
-    pub fn drain_book_feed_with_clock(&mut self, mut wall_clock_ms: impl FnMut() -> i64) {
-        let mut received_at_ms = None;
-        for _ in 0..BOOK_DRAIN_BUDGET {
-            match self.book_events.try_recv() {
-                Ok(event) => {
-                    let received_at_ms = *received_at_ms.get_or_insert_with(&mut wall_clock_ms);
-                    self.tape_mut().handle_depth_event_at(event, received_at_ms);
-                }
-                Err(mpsc::error::TryRecvError::Empty) => break,
-                Err(mpsc::error::TryRecvError::Disconnected) => {
-                    if self.tape().enabled() && !self.book_channel_closed_reported {
-                        tracing::warn!(
-                            target: "quantick::app",
-                            schema_version = 1_u8,
-                            event_code = "HEATMAP_EVENT_CHANNEL_CLOSED",
-                            symbol = self.symbol.as_str(),
-                            action = "retain_last_book_and_wait_for_feed_switch",
-                            "depth event channel closed"
-                        );
-                        self.book_channel_closed_reported = true;
-                    }
-                    break;
-                }
-            }
-        }
-    }
-
-    /// Delay observed when the newest live trade reached the UI.
-    ///
-    /// `None` while a session is replaying: those prints are as old as the day
-    /// they were recorded, so their original arrival latency is unavailable.
-    ///
-    /// This figure freezes between prints — it is an observation, not a
-    /// measurement of now. [`Self::tape_age_ms`] is the one that ages.
-    pub fn trade_arrival_ms(&self) -> Option<i64> {
-        if self.replay.is_some() {
-            return None;
-        }
-        // Never the forced split's figure. This is a *measurement* — it reaches
-        // the log as APP_HIGH_TRADE_LAG, the control feed scope and the health
-        // view, none of which carry a marker saying a capture run invented it,
-        // and inferred data that is not labelled as such is the one thing this
-        // repo does not ship. The hook drives the readout through the latency
-        // port instead, where every consumer already knows the figure is the
-        // feed's own and not the chart's.
-        self.latest_trade_latency_ms
-    }
-
-    /// How old the newest event on the tape is, right now.
-    ///
-    /// Deterministic half: the caller supplies wall clock. Takes the newer of
-    /// the trade stream and the book, so a symbol with depth but a thin tape
-    /// is not called stale while its book is live. `None` while replaying, and
-    /// before anything has arrived — nothing to be stale about yet.
-    pub fn tape_age_at(&self, now_ms: i64) -> Option<i64> {
-        if self.replay.is_some() {
-            return None;
-        }
-        let newest = match (self.latest_trade_ms, self.tape().last_event_ms()) {
-            (Some(trade), Some(book)) => Some(trade.max(book)),
-            (trade, book) => trade.or(book),
-        }?;
-        Some(now_ms.saturating_sub(newest).max(0))
-    }
-
-    /// Data-honesty label for how each print is known, or `None` when the venue
-    /// reports true trades and true sides (§8 — the status bar's middle
-    /// section). The label shares its row with the machinery readouts, so it
-    /// stays short and the full story lives in the hover.
-    ///
-    /// A venue that prints nothing at all takes precedence over how sides were
-    /// decided: on a quote-driven feed *every* print is derived, and saying so
-    /// is the more important disclosure. Without it a chart of one-unit prints
-    /// reads as a market where every trade happened to be the same size.
-    pub fn side_note(&self, config: &AppConfig) -> Option<(String, Option<String>)> {
-        if let Some(link) = &self.replay {
-            Some((
-                match link.session.header.side_source.as_deref() {
-                    Some(source) => format!("side: {source}"),
-                    None => "side: not recorded".to_owned(),
-                },
-                None,
-            ))
-        } else if config.provider_of(&self.active.0).is_some()
-            && !self.capabilities(config).traded_volume
-        {
-            Some((
-                "prints: quote-derived".to_owned(),
-                Some(
-                    "this venue quotes prices but prints no trades: every candle is built \
-                     from one synthetic print per tick, at the mid of bid and ask, carrying \
-                     one unit — never a traded size"
-                        .to_owned(),
-                ),
-            ))
-        } else {
-            // The running feed, not the still-uncommitted selection.
-            config
-                .side_note(&self.active.0)
-                .map(|note| (note.to_owned(), None))
-        }
-    }
-
-    /// Make a recorded session the chart's source, replacing whatever feed is
-    /// running. The live selection is untouched, so closing the replay comes
-    /// back to exactly the feed and symbol that were streaming before.
-    pub fn open_replay(&mut self, config: &AppConfig, request: quantick_feed::ReplayRequest) {
-        tracing::info!(
-            target: "quantick::app",
-            schema_version = 1_u8,
-            event_code = "REPLAY_OPENED",
-            session = %request.session.label(),
-            file = %request.session.path.display(),
-            trades = request.session.trades.len(),
-            speed = request.options.speed,
-            action = "replace_feed_source",
-            "opening a recorded session"
-        );
-
-        // A position on the tape belongs to the session that is ending:
-        // flatten it while the journal still carries that session's
-        // source. attach() flips the source to the new feed's, and a
-        // flatten after it would file the old session's trade under the
-        // new session's name — a live trade laundered into the practice
-        // record. (reset_market_state below flattens again: a no-op.)
-        self.paper.on_timeline_reset();
-        let source = feed::FeedSource::Replay(Box::new(request));
-        let handle = feed::spawn(source, &config.metatrader, shelf_dir());
-        self.attach(handle);
-
-        if let Some(link) = &self.replay {
-            self.symbol = link.symbol().to_string();
-        }
-        self.refresh_chip_label(config);
-        // Depth is not in a recording; the toggle is disabled by capability,
-        // and the view must not keep drawing a book from the live feed.
-        let generation = self.next_book_generation();
-        self.tape_mut().set_enabled(false, generation);
-        self.reset_market_state();
-    }
-
-    /// Leave replay and put the live feed back.
-    pub fn close_replay(&mut self, config: &AppConfig) {
-        if self.replay.take().is_none() {
-            return;
-        }
-        let (feed_id, symbol) = self.active.clone();
-        self.feed_id = feed_id;
-        self.symbol = symbol;
-        self.refresh_chip_label(config);
-        tracing::info!(
-            target: "quantick::app",
-            schema_version = 1_u8,
-            event_code = "REPLAY_CLOSED",
-            feed = %self.feed_id,
-            symbol = %self.symbol,
-            action = "respawn_live_feed",
-            "leaving market replay"
-        );
-
-        let Some(provider) = config.provider_of(&self.feed_id) else {
-            // The configuration changed under us; there is nothing to go back
-            // to, so the chart stays as it is rather than dying.
-            self.reset_market_state();
-            return;
-        };
-        // Same ordering rule as open_replay: the flatten of a replay
-        // position must journal under the replay source, before attach()
-        // flips the journal back to live — or a practice trade counts in
-        // the real track record.
-        self.paper.on_timeline_reset();
-        let handle = feed::spawn_live(provider, &self.symbol, &config.metatrader, shelf_dir());
-        self.attach(handle);
-        self.reset_market_state();
-    }
-
-    /// Respawn the transport and keep everything the chart has built:
-    /// bars, drawings, indicators, armed strategies, the open paper position.
-    ///
-    /// The cheap half of recovery, and the one a trader reaches for when the
-    /// feed hiccuped: nothing on screen moves. Before it existed the only
-    /// button the feed's notice offered said "Try again" and quietly ran
-    /// [`Self::reload_feed`] — which flattens the position and disarms every
-    /// strategy. A control that costs a trader their position without saying so
-    /// is not a recovery control.
-    ///
-    /// The new session replays its own recent window, so the market time the
-    /// chart had already reached becomes a floor
-    /// ([`quantick_feed::past_resume_floor`]): everything at or before it is
-    /// overlap and is dropped, and the first print past it decides whether the
-    /// silence was long enough to be a marked gap. A replay owns the chart
-    /// while it plays and has no transport to recover.
-    /// Returns whether the feed was really respawned. `false` means there was
-    /// nothing to respawn — a recorded session owns the chart, or the tab's
-    /// feed id is no longer in the feed table — and a caller told `true` when
-    /// nothing happened is exactly the inferred-versus-observed lie the
-    /// honesty rule forbids.
-    pub fn reconnect_feed(&mut self, config: &AppConfig) -> bool {
-        if self.replay.is_some() {
-            return false;
-        }
-        let Some(provider) = config.provider_of(&self.feed_id) else {
-            return false;
-        };
-        tracing::info!(
-            target: "quantick::app",
-            schema_version = 1_u8,
-            event_code = "FEED_RECONNECTED_BY_USER",
-            feed = %self.feed_id,
-            symbol = %self.symbol,
-            action = "respawn_feed_keeping_timeline",
-            resume_floor_ms = self.latest_trade_ms,
-            "reconnecting the feed and keeping the timeline"
-        );
-        // Taken before the handle is swapped: it is a fact about the chart, not
-        // about the session, and it has to survive the attach.
-        self.resume_floor_ms = self.latest_trade_ms;
-        let handle = feed::spawn_live(provider, &self.symbol, &config.metatrader, shelf_dir());
-        self.attach_resuming(handle);
-        // The book from the dropped socket is gone with it. Nothing is reset
-        // here on purpose: the new session opens with a complete snapshot, and
-        // a snapshot is exactly what replaces a book wholesale.
-        self.ensure_book_capture(config);
-        true
-    }
-
-    /// Throw the timeline away and rebuild it from zero.
-    ///
-    /// The expensive half, for the case the trader described: MetaTrader itself
-    /// froze, so the bars on screen are not merely behind, they are wrong.
-    /// Reconnecting a socket that is already open fixes nothing there; what
-    /// they did by hand was close quantick and open it again, and this is that
-    /// act without the restart.
-    ///
-    /// It costs what a restart costs, which is why the card says so before it
-    /// is pressed: [`Self::reset_market_state`] disarms every strategy and
-    /// [`PaperTrading::on_timeline_reset`] closes and journals an open
-    /// position. A position cannot honestly survive into a rebuilt timeline.
-    /// Returns whether the chart was really rebuilt; see
-    /// [`Self::reconnect_feed`] for why the answer is reported rather than
-    /// assumed.
-    pub fn reload_feed(&mut self, config: &AppConfig) -> bool {
-        if self.replay.is_some() {
-            return false;
-        }
-        let Some(provider) = config.provider_of(&self.feed_id) else {
-            return false;
-        };
-        tracing::info!(
-            target: "quantick::app",
-            schema_version = 1_u8,
-            event_code = "FEED_RELOADED_BY_USER",
-            feed = %self.feed_id,
-            symbol = %self.symbol,
-            action = "respawn_feed",
-            "rebuilding the chart from a new feed session"
-        );
-        // A rebuild has no timeline to resume onto, so no floor and no seam.
-        self.resume_floor_ms = None;
-        let handle = feed::spawn_live(provider, &self.symbol, &config.metatrader, shelf_dir());
-        self.attach(handle);
-        self.reset_market_state();
-        // The live market is back and it can stream depth again; start
-        // recording immediately rather than waiting for the map to be opened.
-        self.ensure_book_capture(config);
-        true
-    }
-
-    /// Take in a batch from a session that resumed onto a kept timeline,
-    /// keeping only what the chart has not already seen.
-    ///
-    /// The prints that survive reach the bars, the tape and the simulator's
-    /// *mark*, and nothing else — no fill, no strategy, no arrival-latency
-    /// reading. They are history: the market made them minutes ago while
-    /// nobody was listening. Run through the live path they would fill a
-    /// resting limit at a price the trader could never have been filled at,
-    /// fire a strategy on a bar that is already over, and report the length of
-    /// the outage as this feed's delay. `Backfilled` states the same rule in
-    /// its own words — history only seeds the mark — and this is the other
-    /// place history arrives.
-    ///
-    /// The floor lives for exactly one event, whatever that event contained.
-    /// A session replays its window once, at the start; leaving the floor up
-    /// past that swallowed the next *load older* answer whole (every older
-    /// print is below the floor), and on a venue that replays nothing it
-    /// waited for the next natural print and called the quiet in between a
-    /// gap.
-    ///
-    /// Returns whether anything was actually ingested, so the caller's live
-    /// flag means the same thing on this path as on the ordinary one.
-    fn ingest_resumed(&mut self, trades: &[quantick_engine::Trade]) -> bool {
-        let Some(floor) = self.resume_floor_ms.take() else {
-            return false;
-        };
-        let fresh = past_resume_floor(trades, floor);
-        let overlap = trades.len() - fresh.len();
-        if overlap > 0 {
-            // Logged rather than silent, like the same rule inside
-            // `quantick-feed-mt5`: a print that was dropped is a print the
-            // chart chose not to count, and that choice is reviewable.
-            tracing::info!(
-                target: "quantick::app",
-                schema_version = 1_u8,
-                event_code = "FEED_RESUME_OVERLAP_DROPPED",
-                feed = %self.feed_id,
-                symbol = %self.symbol,
-                dropped = overlap,
-                floor_ms = floor,
-                "dropped the window a resumed session replayed"
-            );
-        }
-        let Some(first) = fresh.first() else {
-            return false;
-        };
-        self.record_gap(floor, first.timestamp_ms);
-        self.history_trades += fresh.len();
-        self.latest_trade_ms = Some(fresh[fresh.len() - 1].timestamp_ms);
-        for trade in fresh {
-            for pane in self.panes_mut() {
-                pane.ingest_live_trade(trade);
-            }
-        }
-        // The mark, and only the mark — the same seeding `Backfilled` does.
-        self.paper.seed(&fresh[fresh.len() - 1]);
-        true
-    }
-
-    /// Land the gap `QUANTICK_FEED_GAP` asked for, once the chart has bars for
-    /// it to sit between.
-    ///
-    /// Placed at the open of the bar halfway through the series, so the seam is
-    /// on screen at the zoom a capture opens on rather than off the left edge.
-    /// It goes through [`Self::record_gap`], so a hooked run and a real
-    /// reconnect draw the same mark from the same list.
-    fn land_demo_gap(&mut self) {
-        let Some(silence_ms) = self.pending_demo_gap_ms else {
-            return;
-        };
-        let bars = self.flow_pane.state.bars();
-        // Two bars at least: a seam is a boundary, and a boundary needs a bar
-        // on each side of it.
-        if bars.len() < 2 {
-            return;
-        }
-        let to_ms = bars[bars.len() / 2].open_time;
-        self.pending_demo_gap_ms = None;
-        self.record_gap(to_ms - silence_ms, to_ms);
-    }
-
-    /// Record the silence a reconnect left, when it is long enough to be worth
-    /// marking.
-    ///
-    /// Short silences are the recovery path working and marking them would be
-    /// noise that teaches the trader to stop reading marks; see
-    /// [`MIN_MARKED_GAP_MS`].
-    fn record_gap(&mut self, from_ms: i64, to_ms: i64) {
-        let gap = FeedGap { from_ms, to_ms };
-        if gap.duration_ms() < MIN_MARKED_GAP_MS {
-            return;
-        }
-        // Bounded: the newest gaps are the ones on screen, so the oldest is
-        // what falls off.
-        if self.feed_gaps.len() >= MAX_REMEMBERED_GAPS {
-            self.feed_gaps.remove(0);
-        }
-        self.feed_gaps.push(gap);
-        tracing::info!(
-            target: "quantick::app",
-            schema_version = 1_u8,
-            event_code = "FEED_GAP_MARKED",
-            feed = %self.feed_id,
-            symbol = %self.symbol,
-            from_ms = gap.from_ms,
-            to_ms = gap.to_ms,
-            duration_ms = gap.duration_ms(),
-            "marked a stretch of market time no print covers"
-        );
-    }
-    /// Lay the canvas out and run every visible pane through it (§11).
-    ///
-    /// Single is one pane over the whole area — the same code path the split
-    /// takes, with one pane in it, so the default layout can never drift from
-    /// the split one.
-    pub fn draw_canvas(
-        &mut self,
-        ui: &mut egui::Ui,
-        area: egui::Rect,
-        chrome: &mut CanvasChrome<'_>,
-    ) {
-        // How many context charts this layout asks for, and how many the tab
-        // has actually built. The lower of the two is what gets drawn: a
-        // layout may name a pane the tab is still building, and half a canvas
-        // is better than a frame of nothing.
-        self.last_canvas_width = area.width();
-        let context_wanted = self
-            .layout
-            .kinds()
-            .iter()
-            .filter(|kind| matches!(kind, PaneKind::Time))
-            .count();
-        let context_shown = context_wanted.min(self.time_panes.len());
-        let show_time = context_shown > 0;
-        // The flow pane also stands in for a time pane still being built, so
-        // the frame between asking for the Time layout and the pane existing
-        // shows the market rather than nothing.
-        let show_flow = self.layout.shows_flow() || !show_time;
-        let split = show_time && show_flow;
-        // One visible pane is handed the whole canvas and the rest of this
-        // reduces to nothing: no divider, no focus rule — though a lone time
-        // pane keeps its header.
-        let (time_area, divider, flow_area) = if split {
-            let width = if self.context_collapsed {
-                canvas_layout::PaneWidth::Collapsed {
-                    restore: self.split_fraction,
-                }
-            } else {
-                canvas_layout::PaneWidth::Manual(self.split_fraction)
-            };
-            let row = canvas_layout::split_row(area, &[width, canvas_layout::PaneWidth::Auto]);
-            (Some(row.panes[0]), Some(row.dividers[0]), row.panes[1])
-        } else if show_time {
-            (Some(area), None, area)
-        } else {
-            (None, None, area)
-        };
-        // A collapsed column paints a rail, not charts: eight pixels is a
-        // handle, not a chart, and laying one out there would draw a price
-        // axis and nothing else.
-        let collapsed_rail = (split && self.context_collapsed)
-            .then_some(time_area)
-            .flatten();
-        let time_area = if collapsed_rail.is_some() {
-            None
-        } else {
-            time_area
-        };
-
-        // The context column, carved into one band per chart it shows, top to
-        // bottom. Each band spends its own header strip and hands back the
-        // chart rect below it.
-        let mut context_charts: SmallVec<[egui::Rect; MAX_CONTEXT_PANES]> = SmallVec::new();
-        if let Some(column) = time_area {
-            // Focus before input, so the click that focuses a pane is also the
-            // click that pane goes on to handle. Only a split has focus to
-            // move: a single visible pane is the focused one by definition.
-            let heights: SmallVec<[canvas_layout::PaneWidth; MAX_CONTEXT_PANES]> =
-                SmallVec::from_elem(canvas_layout::PaneWidth::Auto, context_shown);
-            let bands = canvas_layout::split_column(column, &heights);
-            if split {
-                self.focus_from_pointer(ui, &bands.panes[..context_shown], flow_area);
-            }
-            for (slot, band) in bands.panes.iter().enumerate().take(context_shown) {
-                let areas = split_time_pane(*band);
-                // Each context chart carries its own timeframe selector (§11):
-                // its BARS group, beside the toolbar's, which keeps governing
-                // the flow pane.
-                let mut interval_ms = self.time_panes[slot].time_interval_ms;
-                let header_layout = crate::time_header::draw(
-                    ui,
-                    areas.header,
-                    &mut interval_ms,
-                    self.time_panes[slot].id,
-                    &self.time_panes[slot].layout_label,
-                );
-                #[cfg(test)]
-                if slot == 0 {
-                    self.time_header_chips = header_layout.chips();
-                }
-                if header_layout.changed {
-                    let pane = &mut self.time_panes[slot];
-                    pane.kind = BarKind::Time;
-                    pane.time_interval_ms = interval_ms;
-                }
-                context_charts.push(areas.chart);
-            }
-        }
-
-        // Which shared mark the pointer is over, on each pane, against the
-        // other pane's store. Answered here because answering it needs both
-        // panes at once, and the loop below holds them one at a time.
-        let picks = self.shared_picks(ui);
-
-        let mut edits: SmallVec<[(PaneIndex, SharedInteraction); MAX_CANVAS_PANES]> =
-            SmallVec::new();
-        {
-            // Focus as an address, so the loop below compares like with
-            // like however many panes it walks.
-            let focused = self.focused_side().index();
-            let Self {
-                flow_pane,
-                time_panes,
-                symbol,
-                paper,
-                feed_gaps,
-                paper_drag_pane,
-                ..
-            } = self;
-            // Cleared before the loop, set by whichever panes it actually
-            // walks. A collapsed context column draws nothing and would
-            // otherwise keep the rect it had when it was last open, which
-            // `starved_pane` would then offer as somewhere to paint the
-            // offline note — off the visible canvas, on a chart that is not
-            // there.
-            flow_pane.last_area = None;
-            for pane in time_panes.iter_mut() {
-                pane.last_area = None;
-            }
-            // The time pane has no tape of its own (§11), so its footprint
-            // rows adopt the flow pane's capture bucket — the instrument's
-            // grid is a fact about the market, not about which pane shows it.
-            //
-            // Which is why there is no longer a gate here. This used to run
-            // only while the time pane's *footprint layer* was visible, and
-            // that contradicted the sentence above it: the ladders have a
-            // second consumer now, and a fixed-range volume profile folds them
-            // with the layer hidden. So the same profile, on the same market,
-            // read at the flow pane's bucket on one chart and at the default
-            // on the other — a hundredfold difference in row height on WDO,
-            // which paints as a slab beside a wash. Two surfaces that are the
-            // same thing have to behave the same way.
-            //
-            // Unconditional is also cheap: `set_footprint_group` returns
-            // immediately when the bucket has not changed, which is every
-            // frame but the one after a market switch.
-            if let (Some(time), Some(base)) = (
-                time_panes.first_mut(),
-                flow_pane
-                    .orderflow
-                    .as_ref()
-                    .map(|tape| tape.base_capture_grouping()),
-            ) {
-                time.state.set_footprint_group(base);
-            }
-            // Context panes carry addresses `1..`, the flow pane `0` — the
-            // order `Tab::pane_at` uses, never the order they sit in.
-            let addressed: SmallVec<[(PaneIndex, egui::Rect); MAX_CANVAS_PANES]> = context_charts
-                .iter()
-                .enumerate()
-                .map(|(slot, rect)| (slot + 1, *rect))
-                .chain(show_flow.then_some((0 as PaneIndex, flow_area)))
-                .collect();
-            let trading_pane = trading_pane(
-                ui.ctx().pointer_latest_pos(),
-                &addressed,
-                paper.gesture_active(),
-                *paper_drag_pane,
-                focused,
-            );
-            let mut chrome = PaneChrome {
-                toolrail: chrome.toolrail,
-                presets: chrome.presets,
-                begin_text_edit: chrome.begin_text_edit,
-                style: chrome.style,
-                tz: chrome.tz,
-                symbol,
-                paper,
-                paper_takes_input: false,
-                paper_hud_here: false,
-                shared_pick: None,
-                shared: SharedInteraction::default(),
-                feed_gaps: &feed_gaps[..],
-                capabilities: chrome.capabilities,
-                side_inferred: chrome.side_inferred,
-                footprint: chrome.footprint,
-                layers: chrome.layers,
-            };
-            // Time pane first, then flow. Both take the same two steps in the
-            // same order — which is what keeps the split honest: the second
-            // pane cannot drift from the first, and one pane is this same
-            // loop with one entry in it.
-            // Context panes carry addresses `1..`, the flow pane `0` — the
-            // order `Tab::pane_at` uses, never the order they sit in.
-            let context = time_panes
-                .iter_mut()
-                .zip(context_charts.iter().copied())
-                .enumerate()
-                .map(|(slot, (pane, chart))| (pane, chart, slot + 1));
-            let flow = show_flow.then_some((&mut *flow_pane, flow_area, 0 as PaneIndex));
-            for (pane, rect, side) in context.chain(flow) {
-                chrome.paper_takes_input = side == trading_pane;
-                // The HUD is one card and follows focus, so it does not
-                // flicker from pane to pane as the hand crosses them.
-                chrome.paper_hud_here = side == focused;
-                chrome.shared_pick = picks.for_pane(side);
-                chrome.shared = SharedInteraction::default();
-                pane.handle_navigation(ui, rect, &mut chrome);
-                pane.draw_chart(ui.painter(), rect, &mut chrome);
-                // Whatever this pane did to the other's marks travels out of
-                // the loop: the store it belongs to is the pane that is not
-                // borrowed right now.
-                if chrome.shared != SharedInteraction::default() {
-                    edits.push((side, chrome.shared));
-                }
-            }
-            // Written *after* the loop, not before it. The drag begins
-            // inside `handle_chart_input`, so on the frame of the press
-            // `gesture_active()` is still false up there and the pin would
-            // be stored as `None` — leaving the very next frame, the first
-            // one that actually drags, to fall through to the pointer. A
-            // flick across a divider in one frame then repriced the order
-            // against the neighbour's scale, which is the whole thing the
-            // pin exists to stop.
-            *paper_drag_pane = paper.gesture_active().then_some(trading_pane);
-        }
-        self.apply_shared_interactions(&edits);
-
-        // Drawings marked "show on all charts" cross here, after both panes
-        // have drawn and cached their projections. It happens outside the
-        // loop above because each pane paints the *other* pane's marks, and
-        // that needs both panes borrowed at once — immutably, which is also
-        // the guarantee that a foreign mark can only be looked at.
-        self.paint_shared_drawings(ui.painter());
-
-        // The position HUD rides the pane that owns order entry (the focused
-        // one). It draws here, after the pane loop, because its buttons need
-        // the paper host mutably — inside the loop that borrow is pinned
-        // behind the shared chrome.
-        if let Some((rect, scale)) = self.focused_pane().paper_hud_anchor() {
-            crate::paper_hud::draw(ui.ctx(), rect, &mut self.paper, &scale);
-        }
-
-        if let Some(rail) = collapsed_rail {
-            self.draw_collapsed_rail(ui, rail);
-        }
-        let (Some(time_area), Some(divider)) = (time_area, divider) else {
-            return;
-        };
-        self.draw_canvas_divider(ui, divider, area.width());
-        // §11: a 1 px accent under the focused pane's top edge — no border
-        // boxes around market data.
-        let focused = match self.focused_side() {
-            PaneSide::Time(slot) => context_charts.get(slot).copied().unwrap_or(time_area),
-            PaneSide::Flow => flow_area,
-        };
-        ui.painter().line_segment(
-            [
-                egui::pos2(focused.left(), focused.top() + FOCUS_RULE_PX / 2.0),
-                egui::pos2(focused.right(), focused.top() + FOCUS_RULE_PX / 2.0),
-            ],
-            egui::Stroke::new(FOCUS_RULE_PX, theme::ACCENT),
-        );
-    }
-
-    /// What the pointer is over, on each pane, among the *other* pane's
-    /// shared marks.
-    ///
-    /// Resolved with both panes in hand and before either is borrowed for its
-    /// input pass, which is the only moment a pane can be asked about marks it
-    /// does not hold. Nothing to answer on an unsplit tab: one pane has no
-    /// other pane to mirror.
-    fn shared_picks(&self, ui: &egui::Ui) -> SharedPicks {
-        let count = self.pane_count();
-        let mut picks = SharedPicks {
-            by_pane: SmallVec::from_elem(None, count),
-        };
-        if count < 2 {
-            // One pane has no other pane to mirror.
-            return picks;
-        }
-        let Some(position) = ui.input(|input| input.pointer.latest_pos()) else {
-            return picks;
-        };
-
-        for viewer in 0..count {
-            let Some(pane) = self.pane_at(viewer) else {
-                continue;
-            };
-            // Only the pane the pointer is actually over is asked. Besides
-            // saving the work, it is what stops a horizontal line — which
-            // spans a whole chart — from reporting a hit on the pane beside
-            // the one the pointer is in, at the same height.
-            if !pane
-                .last_chart_area
-                .is_some_and(|chart| chart.contains(position))
-            {
-                continue;
-            }
-            // The mark may belong to any other pane. First owner in address
-            // order wins, which is stable frame to frame: a pick that
-            // depended on iteration luck would move an object between charts
-            // between frames.
-            for owner in (0..count).filter(|owner| *owner != viewer) {
-                let Some(source) = self.pane_at(owner) else {
-                    continue;
-                };
-                if let Some((index, anchor)) = pane.shared_pick(source, position) {
-                    picks.by_pane[viewer] = Some(SharedPick {
-                        owner,
-                        index,
-                        anchor,
-                        locked: source.drawings.items()[index].locked,
-                    });
-                    break;
-                }
-            }
-        }
-        picks
-    }
-
-    /// Land what each pane did to the other's marks on the store that holds
-    /// them.
-    ///
-    /// The gesture brackets travel with the edits so a whole drag started on
-    /// the mirror lands as one undo entry on the owning store — the same
-    /// coalescing the object gets on its own chart, because it is the same
-    /// gesture on the same object. A selection taken on one pane is dropped on
-    /// the other, so the tab never holds two.
-    fn apply_shared_interactions(&mut self, edits: &[(PaneIndex, SharedInteraction)]) {
-        for (actor, interaction) in edits {
-            // No owner means no mark was ever taken hold of, so there is
-            // nothing to land. Refused rather than guessed: landing it on a
-            // neighbour chosen by arithmetic would move an object the trader
-            // drew onto a chart they were not working on.
-            let Some(owner) = interaction.owner else {
-                continue;
-            };
-            if self.pane_at(owner).is_none() {
-                continue;
-            }
-            if interaction.begin_gesture {
-                self.pane_at_mut(owner)
-                    .expect("owner checked above")
-                    .drawings
-                    .begin_gesture();
-            }
-            if let Some(edit) = interaction.edit {
-                if matches!(edit, SharedEdit::Select(_))
-                    && let Some(pane) = self.pane_at_mut(*actor)
-                {
-                    // A selection taken on a mirror is dropped on the pane
-                    // that took it, so the tab never holds two.
-                    pane.drawings.select(None);
-                }
-                self.pane_at_mut(owner)
-                    .expect("owner checked above")
-                    .apply_shared_edit(edit);
-            }
-            if interaction.commit_gesture {
-                self.pane_at_mut(owner)
-                    .expect("owner checked above")
-                    .drawings
-                    .commit_gesture();
-            }
-        }
-    }
-
-    /// Cross-pane drawings: each pane paints the shared marks of the other.
-    ///
-    /// Nothing happens on an unsplit tab — one pane has no other pane to
-    /// borrow marks from, and the drawing is already on the only chart there
-    /// is. Scope stops here, at the tab: the panes below hold one symbol on
-    /// one feed, which is what makes a price level mean the same thing on
-    /// both (`docs/ux/drawing-tools-2026-08.md` §D7).
-    fn paint_shared_drawings(&self, painter: &egui::Painter) {
-        let count = self.pane_count();
-        if count < 2 {
-            return;
-        }
-        for viewer in 0..count {
-            let Some(pane) = self.pane_at(viewer) else {
-                continue;
-            };
-            for owner in (0..count).filter(|owner| *owner != viewer) {
-                if let Some(source) = self.pane_at(owner) {
-                    pane.paint_shared_from(painter, source);
-                }
-            }
-        }
-    }
-
-    /// Clicking a pane focuses it (§11). Read from the raw pointer press
-    /// rather than a widget response, so the press that starts a pan or picks
-    /// up a drawing focuses the pane it landed in on that same frame.
-    fn focus_from_pointer(
-        &mut self,
-        ui: &egui::Ui,
-        context_bands: &[egui::Rect],
-        flow_area: egui::Rect,
-    ) {
-        let pressed = ui.input(|input| {
-            input
-                .pointer
-                .primary_pressed()
-                .then(|| input.pointer.interact_pos())
-                .flatten()
-        });
-        let Some(position) = pressed else { return };
-        // A press egui routed to another layer belongs to whatever floats
-        // there — the toast, the object manager, the inspector — not to the
-        // pane it happens to cover. Taking those as pane clicks made the
-        // toast's Undo act on the chart the button floated over rather than
-        // the one it was raised for.
-        if ui.ctx().layer_id_at(position) != Some(ui.layer_id()) {
-            return;
-        }
-        // Each band of the context column is its own pane (§11): the press
-        // names the chart it landed in, not the column. Before this the whole
-        // column was one target, which is how a third pane could never take
-        // focus.
-        if let Some(slot) = context_bands
-            .iter()
-            .position(|band| band.contains(position))
-        {
-            self.focus = PaneSide::Time(slot);
-        } else if flow_area.contains(position) {
-            self.focus = PaneSide::Flow;
-        }
-    }
-
-    /// The collapsed context column: a rail with a grip, and the way back.
-    ///
-    /// A pane dragged to nothing has to leave something behind. Blender's
-    /// manual puts the rule plainly — a hidden region leaves a little arrow to
-    /// click — and the vertical axis in this app already refuses zero for the
-    /// same reason (`indicators::COLLAPSED_PANE_HEIGHT_PX`). Eight pixels of a
-    /// 1920 px canvas is four tenths of one percent: near enough to the "size
-    /// zero" a trader asks for, and not so near that the chart is gone for
-    /// good.
-    ///
-    /// The paint is 8 px wide; the *hit* area is 24 px, reaching into the
-    /// chart beside it where it costs nothing but the pointer's first few
-    /// pixels. A rail that photographed well but could not be hit would be a
-    /// picture of an affordance rather than one.
-    fn draw_collapsed_rail(&mut self, ui: &egui::Ui, rail: egui::Rect) {
-        #[cfg(test)]
-        {
-            self.collapsed_rail = Some(rail);
-        }
-        let painter = ui.painter();
-        painter.rect_filled(rail, egui::Rounding::ZERO, theme::CHROME);
-        // The inner edge, so the rail reads as chrome against the chart rather
-        // than as a stripe the chart happens to start after.
-        painter.line_segment(
-            [
-                egui::pos2(rail.right(), rail.top()),
-                egui::pos2(rail.right(), rail.bottom()),
-            ],
-            egui::Stroke::new(1.0_f32, theme::BORDER),
-        );
-
-        let hit = egui::Rect::from_min_max(
-            rail.min,
-            egui::pos2(
-                rail.left() + canvas_layout::COLLAPSED_HIT_PX.max(rail.width()),
-                rail.bottom(),
-            ),
-        );
-        let response = ui
-            .interact(
-                hit,
-                egui::Id::new(("collapsed_context_rail", self.id)),
-                egui::Sense::click(),
-            )
-            .on_hover_text("show the timeframe charts again");
-        if response.hovered() {
-            ui.ctx().set_cursor_icon(egui::CursorIcon::PointingHand);
-        }
-
-        // The grip: a short bar at the rail's middle, in the colour a reader
-        // already knows means "chrome you can take hold of".
-        let grip_colour = if response.hovered() {
-            theme::ACCENT
-        } else {
-            theme::TEXT_MUTED
-        };
-        let grip = egui::Rect::from_center_size(
-            rail.center(),
-            egui::vec2(RAIL_GRIP_WIDTH_PX, RAIL_GRIP_HEIGHT_PX),
-        );
-        ui.painter()
-            .rect_filled(grip, egui::Rounding::same(1.0), grip_colour);
-
-        if response.clicked() {
-            self.set_context_collapsed(false);
-        }
-    }
-
-    /// The divider between the panes, as a resize handle.
-    ///
-    /// Registered after both panes so it takes the drag that would otherwise
-    /// pan the chart behind its grab area, exactly as the live lane's own
-    /// divider does inside a pane.
-    fn draw_canvas_divider(&mut self, ui: &egui::Ui, divider: egui::Rect, canvas_width: f32) {
-        #[cfg(test)]
-        {
-            self.canvas_divider = Some(divider);
-        }
-        ui.painter()
-            .rect_filled(divider, egui::Rounding::ZERO, theme::BORDER);
-        // Namespaced by tab for the same reason a pane namespaces its own ids
-        // (see [`crate::pane`]): egui keeps drag state per id, so one shared
-        // id would let a drag started on this tab's divider carry on into the
-        // next tab's the moment Ctrl+Tab switches under a held button.
-        let handle = ui.interact(
-            divider.expand2(egui::vec2(CANVAS_DIVIDER_HANDLE_PX, 0.0)),
-            egui::Id::new(("canvas_divider", self.id)),
-            egui::Sense::drag(),
-        );
-        if handle.hovered() || handle.dragged() {
-            ui.ctx().set_cursor_icon(egui::CursorIcon::ResizeHorizontal);
-        }
-        if handle.dragged() && canvas_width > 0.0 {
-            // In pixels, because the gesture is in pixels and the floor is
-            // too. `split_fraction` carries the *asked-for* width rather than
-            // a floored one, so a hand that keeps pushing left keeps
-            // travelling: the splitter floors what it draws, and this is what
-            // makes "drag past the floor to dismiss it" a gesture a hand can
-            // finish rather than one that needs a single impossible frame.
-            let wanted_px = self.split_fraction * canvas_width + handle.drag_delta().x;
-            if wanted_px < canvas_layout::COLLAPSE_AT_PX {
-                // Dismissed, not squeezed. `split_fraction` is left where it
-                // was, so the rail springs back to the width the trader chose
-                // rather than to a default that would discard it.
-                self.set_context_collapsed(true);
-            } else {
-                self.set_context_collapsed(false);
-                self.split_fraction = clamp_pane_fraction(wanted_px / canvas_width);
-            }
-        }
-    }
 }
 
 crate::hooks::declare_hooks!["QUANTICK_PANE_COLLAPSED"];
 
 #[cfg(test)]
-mod shared_routing_tests {
-    use super::*;
-    use crate::state::BarSpec;
-    use quantick_feed as feed;
-    use tokio::sync::mpsc;
-
-    /// Hands each test tab its own trades directory. A tab opens a
-    /// paper-trading ledger, and two tabs pointed at one folder read each
-    /// other's trades — which shows up as unrelated ledger tests failing.
-    static NEXT_TEST_DIR: std::sync::atomic::AtomicU32 = std::sync::atomic::AtomicU32::new(0);
-
-    /// A tab with `context` context panes stacked beside its flow pane.
-    fn tab_with_context_panes(context: usize) -> Tab {
-        let (_evt_tx, evt_rx) = mpsc::channel(8);
-        let (_book_tx, book_rx) = mpsc::channel(8);
-        let (cmd_tx, _cmd_rx) = mpsc::channel(8);
-        let mut tab = Tab::new(
-            0,
-            0,
-            "binance".to_owned(),
-            "BTCUSDT".to_owned(),
-            BarSpec::Tick(50),
-            FeedHandle {
-                events: evt_rx,
-                book_events: book_rx,
-                notices: feed::silent_notices(),
-                capabilities: feed::fixed_capabilities(
-                    crate::config::ProviderKind::Binance.capabilities(),
-                ),
-                latency: feed::unsplit_latency(),
-                commands: cmd_tx,
-                replay: None,
-            },
-            // Its own directory, never the shared temp root: a tab opens a
-            // paper-trading ledger, and pointing every test tab at one folder
-            // makes them read each other's trades.
-            crate::scratch::thread_dir("tab-test").join(
-                NEXT_TEST_DIR
-                    .fetch_add(1, std::sync::atomic::Ordering::Relaxed)
-                    .to_string(),
-            ),
-        );
-        for slot in 0..context {
-            tab.time_panes.push(ChartPane::time(
-                100 + slot as u64,
-                crate::time_header::DEFAULT_INTERVAL_MS,
-            ));
-        }
-        tab
-    }
-
-    /// Collapsing must not spend the width it collapses.
-    ///
-    /// `split_fraction` is the trader's own sizing and the only thing that can
-    /// restore the column. A collapse that overwrote it — with a rail's width,
-    /// or with a default — would hand back a different chart from the one that
-    /// was put away.
-    #[test]
-    fn collapsing_a_column_keeps_the_width_it_springs_back_to() {
-        let mut tab = tab_with_context_panes(1);
-        tab.split_fraction = 0.42;
-
-        tab.context_collapsed = true;
-        assert_eq!(
-            tab.split_fraction, 0.42,
-            "the collapse spent the width it was supposed to remember"
-        );
-
-        // And again: a second collapse must not overwrite it either.
-        tab.context_collapsed = true;
-        assert_eq!(tab.split_fraction, 0.42);
-
-        tab.context_collapsed = false;
-        assert_eq!(
-            tab.split_fraction, 0.42,
-            "the column came back at a width the trader never chose"
-        );
-    }
-
-    /// The flow pane is address `0` and the context stack follows it, whatever
-    /// order they are drawn in. A reader who took this for a left-to-right
-    /// order would mirror every edit, so it is pinned.
-    #[test]
-    fn panes_are_addressed_flow_first_then_the_context_stack() {
-        let tab = tab_with_context_panes(2);
-        assert_eq!(tab.pane_count(), 3);
-        assert_eq!(
-            tab.pane_at(0).map(|pane| pane.id),
-            Some(tab.flow_pane.id),
-            "address 0 is the flow pane"
-        );
-        assert_eq!(tab.pane_at(1).map(|pane| pane.id), Some(100));
-        assert_eq!(tab.pane_at(2).map(|pane| pane.id), Some(101));
-        assert!(tab.pane_at(3).is_none(), "there is no fourth pane");
-    }
-
-    /// A shared mark belongs to the pane whose store holds it, and an edit
-    /// made on a mirror has to land *there* — not on "the other pane".
-    ///
-    /// With two panes those two phrases mean the same thing, which is why the
-    /// routing this replaced (`side.other()`) was correct and why nothing
-    /// caught it losing that meaning. With a stack beside the flow pane there
-    /// is more than one other pane. The owner named here is address 2, which
-    /// is neither the actor nor the actor's single counterpart, so this fails
-    /// against the arithmetic it replaced rather than merely passing beside
-    /// it.
-    #[test]
-    fn a_shared_gesture_opens_on_the_pane_the_interaction_names() {
-        let mut tab = tab_with_context_panes(2);
-        tab.apply_shared_interactions(&[(
-            0,
-            SharedInteraction {
-                owner: Some(2),
-                edit: None,
-                begin_gesture: true,
-                commit_gesture: false,
-            },
-        )]);
-
-        assert!(
-            tab.pane_at(2)
-                .expect("the named pane exists")
-                .drawings
-                .in_gesture(),
-            "the gesture must open on the pane the interaction named"
-        );
-        for bystander in [0usize, 1] {
-            assert!(
-                !tab.pane_at(bystander)
-                    .expect("pane exists")
-                    .drawings
-                    .in_gesture(),
-                "pane {bystander} took a gesture it was never named for"
-            );
-        }
-    }
-
-    /// An interaction with no owner is refused rather than guessed. Landing it
-    /// on a neighbour chosen by arithmetic would move an object the trader
-    /// drew onto a chart they were not working on.
-    #[test]
-    fn an_unowned_interaction_lands_nowhere() {
-        let mut tab = tab_with_context_panes(2);
-        tab.apply_shared_interactions(&[(
-            0,
-            SharedInteraction {
-                owner: None,
-                edit: None,
-                begin_gesture: true,
-                commit_gesture: false,
-            },
-        )]);
-        for pane in 0..tab.pane_count() {
-            assert!(
-                !tab.pane_at(pane)
-                    .expect("pane exists")
-                    .drawings
-                    .in_gesture(),
-                "pane {pane} opened a gesture for an interaction that named no owner"
-            );
-        }
-    }
-
-    /// An owner address past the end of the stack is ignored, not panicked on:
-    /// a saved workspace or a control-plane call can name a pane that has
-    /// since gone.
-    #[test]
-    fn an_owner_that_no_longer_exists_is_ignored() {
-        let mut tab = tab_with_context_panes(1);
-        tab.apply_shared_interactions(&[(
-            0,
-            SharedInteraction {
-                owner: Some(7),
-                edit: None,
-                begin_gesture: true,
-                commit_gesture: false,
-            },
-        )]);
-        for pane in 0..tab.pane_count() {
-            assert!(
-                !tab.pane_at(pane)
-                    .expect("pane exists")
-                    .drawings
-                    .in_gesture()
-            );
-        }
-    }
-}
-
-#[cfg(test)]
-mod move_pane_tests {
-    use super::*;
-    use crate::state::BarSpec;
-    use quantick_feed as feed;
-    use tokio::sync::mpsc;
-
-    static NEXT_DIR: std::sync::atomic::AtomicU32 = std::sync::atomic::AtomicU32::new(1000);
-
-    /// One trade at `ms`, cheap enough to build a block from.
-    fn older_trade(ms: i64) -> quantick_engine::Trade {
-        quantick_engine::Trade {
-            agg_id: ms as u64,
-            timestamp_ms: ms,
-            price: rust_decimal::Decimal::new(1000, 1),
-            quantity: rust_decimal::Decimal::ONE,
-            side: quantick_engine::Side::Buy,
-        }
-    }
-
-    /// A tab whose feed channel is still held, so a test can post events onto
-    /// it and drain them the way a frame does.
-    fn tab_with_feed() -> (Tab, mpsc::Sender<FeedEvent>) {
-        let (evt_tx, evt_rx) = mpsc::channel(64);
-        let (_book_tx, book_rx) = mpsc::channel(8);
-        let (cmd_tx, _cmd_rx) = mpsc::channel(8);
-        let tab = Tab::new(
-            0,
-            0,
-            "binance".to_owned(),
-            "BTCUSDT".to_owned(),
-            BarSpec::Tick(50),
-            FeedHandle {
-                events: evt_rx,
-                book_events: book_rx,
-                notices: feed::silent_notices(),
-                capabilities: feed::fixed_capabilities(
-                    crate::config::ProviderKind::Binance.capabilities(),
-                ),
-                latency: feed::unsplit_latency(),
-                commands: cmd_tx,
-                replay: None,
-            },
-            crate::scratch::thread_dir("opening-test").join(
-                NEXT_DIR
-                    .fetch_add(1, std::sync::atomic::Ordering::Relaxed)
-                    .to_string(),
-            ),
-        );
-        (tab, evt_tx)
-    }
-
-    #[test]
-    fn an_opening_slice_draws_without_answering_the_traders_press() {
-        // The defect the trader-ux review caught, and the reason the opening
-        // block has an event of its own rather than reusing the reply.
-        //
-        // The bridge fills the session in behind the chart in thirty-odd
-        // slices. A trader who presses `+ older` while that is happening has
-        // raised a loading indicator and started a campaign; if a slice
-        // arrived as `HistoryPrepended` it would stop that indicator on
-        // history the trader did not ask for, and hand the campaign a page it
-        // did not fetch — so a run could spend its budget, or declare itself
-        // finished, on tape it never pulled.
-        let (mut tab, feed_tx) = tab_with_feed();
-        // The tab already has its own opening request in flight, so the count
-        // is asserted as a delta rather than against zero: what matters is
-        // that an opening slice answers nothing, whatever else is outstanding.
-        tab.loading.begin(LoadingTask::History);
-        let before = tab.loading.count(LoadingTask::History);
-        assert!(before > 0, "the press raised an indicator to begin with");
-
-        let slice: Vec<_> = (0..8).map(|i| older_trade(1_000 + i)).collect();
-        feed_tx
-            .try_send(FeedEvent::OpeningPrepended {
-                trades: slice.clone(),
-                remaining: Some(4),
-            })
-            .expect("the test channel has room");
-        tab.drain_feed();
-
-        assert_eq!(
-            tab.loading.count(LoadingTask::History),
-            before,
-            "an opening slice must leave the trader's own loading indicator up"
-        );
-        assert_eq!(
-            tab.history_trades,
-            slice.len(),
-            "and it is still charted and counted — it is the trader's morning"
-        );
-        assert_eq!(
-            tab.opening_slices_remaining(),
-            Some(4),
-            "and how much of the session is still arriving is readable, not              only loggable: an operator has to be able to tell a chart that is              still filling from one that has all it will get"
-        );
-    }
-
-    #[test]
-    fn a_page_the_trader_asked_for_still_answers_the_press() {
-        // The other half, so the test above cannot pass by the reply path
-        // having quietly stopped working.
-        let (mut tab, feed_tx) = tab_with_feed();
-        tab.loading.begin(LoadingTask::History);
-        let before = tab.loading.count(LoadingTask::History);
-        feed_tx
-            .try_send(FeedEvent::HistoryPrepended(vec![older_trade(1_000)]))
-            .expect("the test channel has room");
-        tab.drain_feed();
-        assert_eq!(
-            tab.loading.count(LoadingTask::History),
-            before - 1,
-            "a reply settles exactly the one press that asked for it"
-        );
-    }
-
-    fn tab_with(context: usize) -> Tab {
-        let (_evt_tx, evt_rx) = mpsc::channel(8);
-        let (_book_tx, book_rx) = mpsc::channel(8);
-        let (cmd_tx, _cmd_rx) = mpsc::channel(8);
-        let mut tab = Tab::new(
-            0,
-            0,
-            "binance".to_owned(),
-            "BTCUSDT".to_owned(),
-            BarSpec::Tick(50),
-            FeedHandle {
-                events: evt_rx,
-                book_events: book_rx,
-                notices: feed::silent_notices(),
-                capabilities: feed::fixed_capabilities(
-                    crate::config::ProviderKind::Binance.capabilities(),
-                ),
-                latency: feed::unsplit_latency(),
-                commands: cmd_tx,
-                replay: None,
-            },
-            crate::scratch::thread_dir("move-test").join(
-                NEXT_DIR
-                    .fetch_add(1, std::sync::atomic::Ordering::Relaxed)
-                    .to_string(),
-            ),
-        );
-        for slot in 0..context {
-            tab.time_panes.push(ChartPane::time(
-                200 + slot as u64,
-                crate::time_header::DEFAULT_INTERVAL_MS,
-            ));
-        }
-        tab
-    }
-
-    /// The order the stack draws in is the order it holds, so moving a chart
-    /// moves the pane rather than swapping what is inside two of them: the
-    /// drawings, indicators and bars travel with the chart the trader moved.
-    #[test]
-    fn moving_a_chart_carries_the_pane_rather_than_its_contents() {
-        let mut tab = tab_with(3);
-        let ids: Vec<u64> = tab.time_panes.iter().map(|pane| pane.id).collect();
-        assert_eq!(ids, vec![200, 201, 202]);
-
-        assert!(
-            tab.move_context_pane(3, 1),
-            "the bottom chart moves to the top"
-        );
-        let after: Vec<u64> = tab.time_panes.iter().map(|pane| pane.id).collect();
-        assert_eq!(
-            after,
-            vec![202, 200, 201],
-            "the pane moved and the others closed up behind it"
-        );
-    }
-
-    #[test]
-    fn moving_a_chart_one_slot_swaps_it_with_its_neighbour() {
-        let mut tab = tab_with(2);
-        assert!(tab.move_context_pane(1, 2));
-        let after: Vec<u64> = tab.time_panes.iter().map(|pane| pane.id).collect();
-        assert_eq!(after, vec![201, 200]);
-    }
-
-    /// The flow pane is address `0` and does not move: its column is the one
-    /// thing every preset agrees on, and a caller that asked to move the
-    /// heatmap meant something this cannot do.
-    #[test]
-    fn the_flow_pane_refuses_to_move() {
-        let mut tab = tab_with(2);
-        let before: Vec<u64> = tab.time_panes.iter().map(|pane| pane.id).collect();
-        assert!(!tab.move_context_pane(0, 1), "address 0 is the flow pane");
-        assert!(!tab.move_context_pane(1, 0), "and it is not a destination");
-        assert_eq!(
-            tab.time_panes
-                .iter()
-                .map(|pane| pane.id)
-                .collect::<Vec<_>>(),
-            before,
-            "a refused move must leave the stack exactly as it was"
-        );
-    }
-
-    /// An address past the end is refused rather than clamped. A control-plane
-    /// call or a stale menu naming a chart that has gone means something this
-    /// cannot do, and moving a different chart would be worse than saying no.
-    #[test]
-    fn an_address_the_stack_does_not_have_is_refused() {
-        let mut tab = tab_with(2);
-        let before: Vec<u64> = tab.time_panes.iter().map(|pane| pane.id).collect();
-        for (from, to) in [(1_usize, 9_usize), (9, 1), (7, 8)] {
-            assert!(
-                !tab.move_context_pane(from, to),
-                "moving {from} to {to} names a chart that is not there"
-            );
-        }
-        assert_eq!(
-            tab.time_panes
-                .iter()
-                .map(|pane| pane.id)
-                .collect::<Vec<_>>(),
-            before
-        );
-    }
-
-    /// Moving a chart onto itself changes nothing, and says so. A caller that
-    /// retried a dropped call needs "nothing happened" to be distinguishable
-    /// from "it worked".
-    #[test]
-    fn moving_a_chart_onto_itself_reports_no_change() {
-        let mut tab = tab_with(2);
-        assert!(!tab.move_context_pane(1, 1));
-        assert!(!tab.move_context_pane(2, 2));
-    }
-
-    /// A single context chart has nowhere to go.
-    #[test]
-    fn a_lone_context_chart_cannot_be_reordered() {
-        let mut tab = tab_with(1);
-        assert!(!tab.move_context_pane(1, 1));
-        assert!(!tab.move_context_pane(1, 2));
-        assert_eq!(tab.time_panes.len(), 1);
-    }
-}
-
-#[cfg(test)]
-mod collapse_path_tests {
-    use super::*;
-    use crate::state::BarSpec;
-    use quantick_feed as feed;
-    use tokio::sync::mpsc;
-
-    static NEXT_DIR: std::sync::atomic::AtomicU32 = std::sync::atomic::AtomicU32::new(2000);
-
-    fn tab() -> Tab {
-        let (_evt_tx, evt_rx) = mpsc::channel(8);
-        let (_book_tx, book_rx) = mpsc::channel(8);
-        let (cmd_tx, _cmd_rx) = mpsc::channel(8);
-        Tab::new(
-            0,
-            0,
-            "binance".to_owned(),
-            "BTCUSDT".to_owned(),
-            BarSpec::Tick(50),
-            FeedHandle {
-                events: evt_rx,
-                book_events: book_rx,
-                notices: feed::silent_notices(),
-                capabilities: feed::fixed_capabilities(
-                    crate::config::ProviderKind::Binance.capabilities(),
-                ),
-                latency: feed::unsplit_latency(),
-                commands: cmd_tx,
-                replay: None,
-            },
-            crate::scratch::thread_dir("collapse-test").join(
-                NEXT_DIR
-                    .fetch_add(1, std::sync::atomic::Ordering::Relaxed)
-                    .to_string(),
-            ),
-        )
-    }
-
-    /// Collapse reports whether it changed anything, so a caller that retried
-    /// a dropped call can tell "nothing happened" from "it worked".
-    #[test]
-    fn collapsing_reports_change_and_is_idempotent() {
-        let mut tab = tab();
-        assert!(
-            tab.set_context_collapsed(true),
-            "the first collapse changes it"
-        );
-        assert!(
-            !tab.set_context_collapsed(true),
-            "the second is a no-op and says so"
-        );
-        assert!(tab.set_context_collapsed(false), "and it comes back");
-        assert!(!tab.set_context_collapsed(false));
-    }
-
-    /// The width the column springs back to is never spent by putting it away
-    /// — however many times it is put away.
-    #[test]
-    fn collapsing_never_spends_the_remembered_width() {
-        let mut tab = tab();
-        tab.split_fraction = 0.42;
-        for _ in 0..3 {
-            tab.set_context_collapsed(true);
-            tab.set_context_collapsed(false);
-        }
-        assert_eq!(
-            tab.split_fraction, 0.42,
-            "the column came back at a width the trader never chose"
-        );
-    }
-
-    /// A workspace saved with its charts put away reopens with them put away.
-    ///
-    /// The order inside `restore_canvas` is the whole test: `set_layout` opens
-    /// the column it reveals, which is right for a menu click and wrong for a
-    /// restore. Assigned before that call, the flag was overwritten every
-    /// time — and the next `capture_arrangement` wrote the wrong answer back
-    /// A three-pane workspace records one bar rule per context chart, and
-    /// each chart opens on its own. Before `context_bars` existed the file
-    /// kept one interval and both charts opened on it — the bottom chart lost
-    /// its timeframe on every restart.
-    #[test]
-    fn a_restored_stack_opens_each_context_chart_on_its_own_interval() {
-        let mut tab = tab();
-        tab.restore_canvas(
-            CanvasLayout::TimeTimeAndFlow,
-            None,
-            false,
-            Some(PaneSide::Time(1)),
-            &[60_000, 900_000],
-            LegendFold::default(),
-        );
-        let config = crate::config::AppConfig {
-            default_feed: "binance".to_owned(),
-            default_symbol: "BTCUSDT".to_owned(),
-            feeds: vec![],
-            metatrader: Default::default(),
-            paper: Default::default(),
-            history: Default::default(),
-        };
-        let style = crate::style::ChartStyle::default();
-        let mut ids = PaneIdAllocator::new();
-        tab.apply_pending_layout(&config, &style, &mut ids);
-        tab.apply_pending_layout(&config, &style, &mut ids);
-        assert_eq!(tab.time_panes.len(), 2, "both charts were built");
-        assert_eq!(tab.time_panes[0].time_interval_ms, 60_000);
-        assert_eq!(
-            tab.time_panes[1].time_interval_ms, 900_000,
-            "the bottom chart opened on the interval the file kept for it"
-        );
-        assert_eq!(
-            tab.focused_side(),
-            PaneSide::Time(1),
-            "and the saved focus names the bottom chart"
-        );
-    }
-
-    /// over the trader's file.
-    #[test]
-    fn a_restored_workspace_keeps_its_collapsed_column() {
-        let mut tab = tab();
-        tab.restore_canvas(
-            CanvasLayout::TimeAndFlow,
-            Some(0.42),
-            true,
-            Some(PaneSide::Flow),
-            &[],
-            LegendFold {
-                flow: false,
-                time: false,
-            },
-        );
-        assert!(
-            tab.context_collapsed,
-            "the workspace recorded a collapsed column and it opened expanded"
-        );
-        assert_eq!(
-            tab.split_fraction, 0.42,
-            "and the width it springs back to is the saved one"
-        );
-    }
-
-    /// Picking the arrangement that is already showing brings the column back.
-    ///
-    /// The picker lights a cell for the current layout; a trader who put the
-    /// column away and then clicked that lit cell is asking for the charts the
-    /// thumbnail draws. Answering with the 8 px rail is the chrome lying.
-    #[test]
-    fn re_picking_the_current_layout_brings_the_column_back() {
-        let mut tab = tab();
-        tab.set_layout(CanvasLayout::TimeAndFlow);
-        assert!(tab.set_context_collapsed(true));
-
-        tab.set_layout(CanvasLayout::TimeAndFlow);
-        assert!(
-            !tab.context_collapsed,
-            "the lit cell promised two panes and handed back a rail"
-        );
-    }
-
-    /// A layout with no flow pane has no column to put away, so the collapse
-    /// flag must not decide its focus.
-    ///
-    /// Read the other way round, `Ctrl+0` on the Timeframe layout pointed
-    /// focus at a pane nobody draws: `paper_hud_here` went false for the one
-    /// chart on screen, so order entry, the ladder and the trade HUD all went
-    /// dead — and with no split there is no rail to click to undo it.
-    #[test]
-    fn collapsing_never_takes_focus_off_the_only_chart_on_screen() {
-        let mut tab = tab();
-        tab.time_panes.push(ChartPane::time(
-            300,
-            crate::time_header::DEFAULT_INTERVAL_MS,
-        ));
-        tab.layout = CanvasLayout::Time;
-        tab.context_collapsed = true;
-
-        assert_eq!(
-            tab.focused_side(),
-            PaneSide::Time(0),
-            "the only chart drawn has to be the one the chrome speaks for"
-        );
-    }
-}
-
-/// Which pane the order-entry gesture lands on, per pointer and per drag.
-#[cfg(test)]
-mod trading_pane_tests {
-    use super::*;
-
-    /// Every visible pane is a trading surface: the aim follows the pointer,
-    /// so holding the buy modifier over a pane that is *not* focused places
-    /// there — no focusing click first. This is the whole of "trade on any
-    /// chart that is on screen".
-    #[test]
-    fn order_entry_follows_the_pointer_onto_an_unfocused_pane() {
-        let top = egui::Rect::from_min_max(egui::pos2(0.0, 0.0), egui::pos2(800.0, 200.0));
-        let bottom = egui::Rect::from_min_max(egui::pos2(0.0, 200.0), egui::pos2(800.0, 400.0));
-        // Pane 1 is the context chart on top, pane 0 the flow chart below.
-        let panes = [(1 as PaneIndex, top), (0 as PaneIndex, bottom)];
-        let focused: PaneIndex = 0;
-
-        assert_eq!(
-            trading_pane(Some(top.center()), &panes, false, None, focused),
-            1,
-            "the pointer is on the unfocused context pane, so the aim is there"
-        );
-        assert_eq!(
-            trading_pane(Some(bottom.center()), &panes, false, None, focused),
-            0,
-            "and on the flow pane it is there"
-        );
-    }
-
-    /// With the pointer nowhere near a pane — over the dock, off the window
-    /// — the focused pane answers. That is also the unsplit case: one pane,
-    /// always the answer, nothing about today changes.
-    #[test]
-    fn order_entry_falls_back_to_focus_with_the_pointer_off_the_canvas() {
-        let flow = egui::Rect::from_min_max(egui::pos2(0.0, 0.0), egui::pos2(800.0, 400.0));
-        let panes = [(0 as PaneIndex, flow)];
-        assert_eq!(
-            trading_pane(Some(egui::pos2(900.0, 200.0)), &panes, false, None, 0),
-            0,
-            "a pointer over the dock leaves entry where focus is"
-        );
-        assert_eq!(
-            trading_pane(None, &panes, false, None, 0),
-            0,
-            "and so does no pointer at all"
-        );
-    }
-
-    /// A drag keeps the pane it started in. The grabbed line is read against
-    /// that pane's price scale; handing the gesture to a neighbour halfway
-    /// through would reprice the order to a different scale — a stop that
-    /// jumps because the hand strayed across a divider.
-    #[test]
-    fn a_paper_drag_stays_with_the_pane_it_started_in() {
-        let top = egui::Rect::from_min_max(egui::pos2(0.0, 0.0), egui::pos2(800.0, 200.0));
-        let bottom = egui::Rect::from_min_max(egui::pos2(0.0, 200.0), egui::pos2(800.0, 400.0));
-        let panes = [(1 as PaneIndex, top), (0 as PaneIndex, bottom)];
-
-        // The drag began on pane 1; the pointer has since crossed into 0.
-        assert_eq!(
-            trading_pane(Some(bottom.center()), &panes, true, Some(1), 0),
-            1,
-            "the gesture stays where it started"
-        );
-        // Released, the pointer decides again immediately.
-        assert_eq!(
-            trading_pane(Some(bottom.center()), &panes, false, Some(1), 0),
-            0,
-            "and the pin is spent the moment the drag ends"
-        );
-    }
-}
+mod tests;
