+++ crates/guards/size-baseline.txt
@@ -170,7 +170,7 @@
 # What is left in `app.rs` is now the window's definition and nothing else:
 # the struct, `new`, `new_with_workspace`, `attach_surface`, the `eframe::App`
 # impl and `mod tests`. The 56 fields are the next cut's job.
-!budget 50996
+!budget 48185
 
 crates/app/src/app.rs 789
 
@@ -277,7 +277,7 @@ crates/app/src/pane.rs 5618
 # which is deliberately not the reply path with a flag -- the two differ in
 # the two lines that matter, and a branch inside the reply arm is how a
 # slice ends up answering a press the trader made.
-crates/app/src/tab.rs 4472
+crates/app/src/tab.rs 1661
 crates/app/src/control/gateway.rs 4142
 crates/app/src/orderflow_render.rs 3075
 crates/app/src/orderflow_view.rs 2485
